Amazon AWS cloud ranges

Retrieves the official list of Amazon cloud network ranges.

amazon_ranges()

Value

list of 4 slots the most interesting of which is prefixes which is a tibble with three columns, the most interesting of which is ip_prefix, but the other columns are useful for various tasks. There is an IPv6 counterpart slot ipv6_prefixes with a similar sub-structure.

Details

Amazon provides their official netblock list in https://ip-ranges.amazonaws.com/ip-ranges.json.

It is unlikely that this list will change in your analysis session, so it is recommended that you cache the results. Future version will automatically cache the results both in-session and on-disk for a period of time.

Examples

ranges <- amazon_ranges()

normalize_ipv4(ranges)

#>   [1] "13.32.0.0/15"      "13.54.0.0/15"      "23.20.0.0/14"     
#>   [4] "27.0.0.0/22"       "34.192.0.0/12"     "35.154.0.0/16"    
#>   [7] "35.156.0.0/14"     "35.160.0.0/13"     "43.250.192.0/24"  
#>  [10] "43.250.193.0/24"   "46.51.128.0/18"    "46.51.192.0/20"   
#>  [13] "46.51.216.0/21"    "46.51.224.0/19"    "46.137.0.0/17"    
#>  [16] "46.137.128.0/18"   "46.137.192.0/19"   "46.137.224.0/19"  
#>  [19] "50.16.0.0/15"      "50.18.0.0/16"      "50.19.0.0/16"     
#>  [22] "50.112.0.0/16"     "52.0.0.0/15"       "52.2.0.0/15"      
#>  [25] "52.4.0.0/14"       "52.8.0.0/16"       "52.9.0.0/16"      
#>  [28] "52.10.0.0/15"      "52.12.0.0/15"      "52.15.0.0/16"     
#>  [31] "52.16.0.0/15"      "52.18.0.0/15"      "52.20.0.0/14"     
#>  [34] "52.24.0.0/14"      "52.28.0.0/16"      "52.29.0.0/16"     
#>  [37] "52.30.0.0/15"      "52.32.0.0/14"      "52.36.0.0/14"     
#>  [40] "52.40.0.0/14"      "52.44.0.0/15"      "52.46.0.0/18"     
#>  [43] "52.48.0.0/14"      "52.52.0.0/15"      "52.54.0.0/15"     
#>  [46] "52.57.0.0/16"      "52.58.0.0/15"      "52.62.0.0/15"     
#>  [49] "52.64.0.0/17"      "52.64.128.0/17"    "52.65.0.0/16"     
#>  [52] "52.66.0.0/16"      "52.67.0.0/16"      "52.68.0.0/15"     
#>  [55] "52.70.0.0/15"      "52.72.0.0/15"      "52.74.0.0/16"     
#>  [58] "52.76.0.0/17"      "52.76.128.0/17"    "52.77.0.0/16"     
#>  [61] "52.78.0.0/16"      "52.79.0.0/16"      "52.80.0.0/16"     
#>  [64] "52.84.0.0/15"      "52.86.0.0/15"      "52.88.0.0/15"     
#>  [67] "52.90.0.0/15"      "52.92.0.0/20"      "52.92.16.0/20"    
#>  [70] "52.92.32.0/22"     "52.92.39.0/24"     "52.92.40.0/21"    
#>  [73] "52.92.48.0/22"     "52.92.52.0/22"     "52.92.56.0/22"    
#>  [76] "52.92.60.0/22"     "52.92.64.0/22"     "52.92.68.0/22"    
#>  [79] "52.92.72.0/22"     "52.92.76.0/22"     "52.92.80.0/22"    
#>  [82] "52.92.92.0/22"     "52.92.248.0/22"    "52.92.252.0/22"   
#>  [85] "52.93.0.0/24"      "52.93.1.0/24"      "52.93.2.0/24"     
#>  [88] "52.93.3.0/24"      "52.93.4.0/24"      "52.93.8.0/22"     
#>  [91] "52.93.12.0/22"     "52.93.16.0/24"     "52.94.0.0/22"     
#>  [94] "52.94.4.0/24"      "52.94.5.0/24"      "52.94.6.0/24"     
#>  [97] "52.94.7.0/24"      "52.94.8.0/24"      "52.94.9.0/24"     
#> [100] "52.94.10.0/24"     "52.94.11.0/24"     "52.94.12.0/24"    
#> [103] "52.94.13.0/24"     "52.94.17.0/24"     "52.94.24.0/23"    
#> [106] "52.94.28.0/23"     "52.94.192.0/22"    "52.94.196.0/24"   
#> [109] "52.94.197.0/24"    "52.94.198.0/28"    "52.94.198.16/28"  
#> [112] "52.94.198.32/28"   "52.94.198.48/28"   "52.94.198.64/28"  
#> [115] "52.94.198.80/28"   "52.94.198.96/28"   "52.94.198.112/28" 
#> [118] "52.94.199.0/24"    "52.94.200.0/24"    "52.94.204.0/23"   
#> [121] "52.94.206.0/23"    "52.94.208.0/21"    "52.94.216.0/21"   
#> [124] "52.94.224.0/20"    "52.94.252.0/23"    "52.94.254.0/23"   
#> [127] "52.95.0.0/20"      "52.95.16.0/21"     "52.95.24.0/22"    
#> [130] "52.95.28.0/24"     "52.95.30.0/23"     "52.95.34.0/24"    
#> [133] "52.95.35.0/24"     "52.95.36.0/22"     "52.95.40.0/24"    
#> [136] "52.95.48.0/22"     "52.95.52.0/22"     "52.95.56.0/22"    
#> [139] "52.95.60.0/24"     "52.95.61.0/24"     "52.95.62.0/24"    
#> [142] "52.95.63.0/24"     "52.95.64.0/20"     "52.95.80.0/20"    
#> [145] "52.95.96.0/22"     "52.95.100.0/22"    "52.95.104.0/22"   
#> [148] "52.95.110.0/24"    "52.95.128.0/22"    "52.95.132.0/22"   
#> [151] "52.95.136.0/23"    "52.95.138.0/24"    "52.95.142.0/23"   
#> [154] "52.95.144.0/24"    "52.95.192.0/20"    "52.95.212.0/22"   
#> [157] "52.95.240.0/24"    "52.95.241.0/24"    "52.95.242.0/24"   
#> [160] "52.95.243.0/24"    "52.95.244.0/24"    "52.95.245.0/24"   
#> [163] "52.95.246.0/24"    "52.95.247.0/24"    "52.95.248.0/24"   
#> [166] "52.95.249.0/24"    "52.95.251.0/24"    "52.95.252.0/24"   
#> [169] "52.95.255.0/28"    "52.95.255.16/28"   "52.95.255.32/28"  
#> [172] "52.95.255.48/28"   "52.95.255.64/28"   "52.95.255.80/28"  
#> [175] "52.95.255.96/28"   "52.95.255.112/28"  "52.95.255.128/28" 
#> [178] "52.95.255.144/28"  "52.192.0.0/15"     "52.196.0.0/14"    
#> [181] "52.200.0.0/13"     "52.208.0.0/13"     "52.216.0.0/15"    
#> [184] "52.218.0.0/18"     "52.218.64.0/18"    "52.218.128.0/18"  
#> [187] "52.218.192.0/18"   "52.219.0.0/20"     "52.219.16.0/22"   
#> [190] "52.219.20.0/22"    "52.219.24.0/21"    "52.219.32.0/21"   
#> [193] "52.219.40.0/22"    "52.219.44.0/22"    "52.219.56.0/22"   
#> [196] "52.219.60.0/23"    "52.219.62.0/23"    "52.219.64.0/22"   
#> [199] "52.219.68.0/22"    "52.219.72.0/22"    "52.219.76.0/22"   
#> [202] "52.219.80.0/20"    "52.220.0.0/15"     "52.222.0.0/17"    
#> [205] "52.222.128.0/17"   "54.64.0.0/15"      "54.66.0.0/16"     
#> [208] "54.67.0.0/16"      "54.68.0.0/14"      "54.72.0.0/15"     
#> [211] "54.74.0.0/15"      "54.76.0.0/15"      "54.78.0.0/16"     
#> [214] "54.79.0.0/16"      "54.80.0.0/13"      "54.88.0.0/14"     
#> [217] "54.92.0.0/17"      "54.92.128.0/17"    "54.93.0.0/16"     
#> [220] "54.94.0.0/16"      "54.95.0.0/16"      "54.144.0.0/14"    
#> [223] "54.148.0.0/15"     "54.150.0.0/16"     "54.151.0.0/17"    
#> [226] "54.151.128.0/17"   "54.152.0.0/16"     "54.153.0.0/17"    
#> [229] "54.153.128.0/17"   "54.154.0.0/16"     "54.155.0.0/16"    
#> [232] "54.156.0.0/14"     "54.160.0.0/13"     "54.168.0.0/16"    
#> [235] "54.169.0.0/16"     "54.170.0.0/15"     "54.172.0.0/15"    
#> [238] "54.174.0.0/15"     "54.176.0.0/15"     "54.178.0.0/16"    
#> [241] "54.179.0.0/16"     "54.182.0.0/16"     "54.183.0.0/16"    
#> [244] "54.184.0.0/13"     "54.192.0.0/16"     "54.193.0.0/16"    
#> [247] "54.194.0.0/15"     "54.196.0.0/15"     "54.198.0.0/16"    
#> [250] "54.199.0.0/16"     "54.200.0.0/15"     "54.202.0.0/15"    
#> [253] "54.204.0.0/15"     "54.206.0.0/16"     "54.207.0.0/16"    
#> [256] "54.208.0.0/15"     "54.210.0.0/15"     "54.212.0.0/15"    
#> [259] "54.214.0.0/16"     "54.215.0.0/16"     "54.216.0.0/15"    
#> [262] "54.218.0.0/16"     "54.219.0.0/16"     "54.220.0.0/16"    
#> [265] "54.221.0.0/16"     "54.222.0.0/19"     "54.222.58.0/28"   
#> [268] "54.222.128.0/17"   "54.223.0.0/16"     "54.224.0.0/15"    
#> [271] "54.226.0.0/15"     "54.228.0.0/16"     "54.229.0.0/16"    
#> [274] "54.230.0.0/16"     "54.231.0.0/17"     "54.231.128.0/19"  
#> [277] "54.231.160.0/19"   "54.231.192.0/20"   "54.231.224.0/21"  
#> [280] "54.231.232.0/21"   "54.231.240.0/22"   "54.231.244.0/22"  
#> [283] "54.231.248.0/22"   "54.231.252.0/24"   "54.231.253.0/24"  
#> [286] "54.231.254.0/24"   "54.232.0.0/16"     "54.233.0.0/18"    
#> [289] "54.233.64.0/18"    "54.233.128.0/17"   "54.234.0.0/15"    
#> [292] "54.236.0.0/15"     "54.238.0.0/16"     "54.239.2.0/23"    
#> [295] "54.239.4.0/22"     "54.239.8.0/21"     "54.239.16.0/20"   
#> [298] "54.239.32.0/21"    "54.239.48.0/22"    "54.239.52.0/23"   
#> [301] "54.239.54.0/23"    "54.239.56.0/21"    "54.239.64.0/21"   
#> [304] "54.239.96.0/24"    "54.239.98.0/24"    "54.239.99.0/24"   
#> [307] "54.239.100.0/23"   "54.239.104.0/23"   "54.239.108.0/22"  
#> [310] "54.239.114.0/24"   "54.239.116.0/22"   "54.239.120.0/21"  
#> [313] "54.239.128.0/18"   "54.239.192.0/19"   "54.240.128.0/18"  
#> [316] "54.240.192.0/22"   "54.240.196.0/24"   "54.240.197.0/24"  
#> [319] "54.240.198.0/24"   "54.240.199.0/24"   "54.240.200.0/24"  
#> [322] "54.240.202.0/24"   "54.240.203.0/24"   "54.240.204.0/22"  
#> [325] "54.240.208.0/22"   "54.240.212.0/22"   "54.240.216.0/22"  
#> [328] "54.240.220.0/22"   "54.240.225.0/24"   "54.240.226.0/24"  
#> [331] "54.240.227.0/24"   "54.240.228.0/23"   "54.240.230.0/23"  
#> [334] "54.240.232.0/22"   "54.240.236.0/22"   "54.240.240.0/24"  
#> [337] "54.240.244.0/22"   "54.240.248.0/21"   "54.241.0.0/16"    
#> [340] "54.242.0.0/15"     "54.244.0.0/16"     "54.245.0.0/16"    
#> [343] "54.246.0.0/16"     "54.247.0.0/16"     "54.248.0.0/15"    
#> [346] "54.250.0.0/16"     "54.251.0.0/16"     "54.252.0.0/16"    
#> [349] "54.253.0.0/16"     "54.254.0.0/16"     "54.255.0.0/16"    
#> [352] "67.202.0.0/18"     "72.21.192.0/19"    "72.44.32.0/19"    
#> [355] "75.101.128.0/17"   "79.125.0.0/17"     "87.238.80.0/21"   
#> [358] "96.127.0.0/17"     "103.4.8.0/22"      "103.8.172.0/22"   
#> [361] "103.246.148.0/23"  "103.246.150.0/23"  "107.20.0.0/14"    
#> [364] "122.248.192.0/18"  "172.96.97.0/24"    "174.129.0.0/16"   
#> [367] "175.41.128.0/18"   "175.41.192.0/18"   "176.32.64.0/19"   
#> [370] "176.32.96.0/21"    "176.32.104.0/21"   "176.32.112.0/21"  
#> [373] "176.32.120.0/22"   "176.32.125.0/25"   "176.34.0.0/19"    
#> [376] "176.34.32.0/19"    "176.34.64.0/18"    "176.34.128.0/17"  
#> [379] "177.71.128.0/17"   "177.72.240.0/21"   "178.236.0.0/20"   
#> [382] "184.72.0.0/18"     "184.72.64.0/18"    "184.72.128.0/17"  
#> [385] "184.73.0.0/16"     "184.169.128.0/17"  "185.48.120.0/22"  
#> [388] "185.143.16.0/24"   "203.83.220.0/22"   "204.236.128.0/18" 
#> [391] "204.236.192.0/18"  "204.246.160.0/22"  "204.246.164.0/22" 
#> [394] "204.246.168.0/22"  "204.246.174.0/23"  "204.246.176.0/20" 
#> [397] "205.251.192.0/19"  "205.251.224.0/22"  "205.251.228.0/22" 
#> [400] "205.251.232.0/22"  "205.251.236.0/22"  "205.251.240.0/22" 
#> [403] "205.251.244.0/23"  "205.251.247.0/24"  "205.251.248.0/24" 
#> [406] "205.251.249.0/24"  "205.251.250.0/23"  "205.251.252.0/23" 
#> [409] "205.251.254.0/24"  "205.251.255.0/24"  "207.171.160.0/20" 
#> [412] "207.171.176.0/20"  "216.137.32.0/19"   "216.182.224.0/20" 
#> [415] "13.54.0.0/15"      "23.20.0.0/14"      "34.192.0.0/12"    
#> [418] "35.154.0.0/16"     "35.156.0.0/14"     "35.160.0.0/13"    
#> [421] "46.51.128.0/18"    "46.51.192.0/20"    "46.51.216.0/21"   
#> [424] "46.51.224.0/19"    "46.137.0.0/17"     "46.137.128.0/18"  
#> [427] "46.137.192.0/19"   "46.137.224.0/19"   "50.16.0.0/15"     
#> [430] "50.18.0.0/16"      "50.19.0.0/16"      "50.112.0.0/16"    
#> [433] "52.0.0.0/15"       "52.2.0.0/15"       "52.4.0.0/14"      
#> [436] "52.8.0.0/16"       "52.9.0.0/16"       "52.10.0.0/15"     
#> [439] "52.12.0.0/15"      "52.15.0.0/16"      "52.16.0.0/15"     
#> [442] "52.18.0.0/15"      "52.20.0.0/14"      "52.24.0.0/14"     
#> [445] "52.28.0.0/16"      "52.29.0.0/16"      "52.30.0.0/15"     
#> [448] "52.32.0.0/14"      "52.36.0.0/14"      "52.40.0.0/14"     
#> [451] "52.44.0.0/15"      "52.48.0.0/14"      "52.52.0.0/15"     
#> [454] "52.54.0.0/15"      "52.57.0.0/16"      "52.58.0.0/15"     
#> [457] "52.62.0.0/15"      "52.64.0.0/17"      "52.64.128.0/17"   
#> [460] "52.65.0.0/16"      "52.66.0.0/16"      "52.67.0.0/16"     
#> [463] "52.68.0.0/15"      "52.70.0.0/15"      "52.72.0.0/15"     
#> [466] "52.74.0.0/16"      "52.76.0.0/17"      "52.76.128.0/17"   
#> [469] "52.77.0.0/16"      "52.78.0.0/16"      "52.79.0.0/16"     
#> [472] "52.80.0.0/16"      "52.86.0.0/15"      "52.88.0.0/15"     
#> [475] "52.90.0.0/15"      "52.95.240.0/24"    "52.95.241.0/24"   
#> [478] "52.95.242.0/24"    "52.95.243.0/24"    "52.95.244.0/24"   
#> [481] "52.95.245.0/24"    "52.95.246.0/24"    "52.95.247.0/24"   
#> [484] "52.95.248.0/24"    "52.95.249.0/24"    "52.95.251.0/24"   
#> [487] "52.95.252.0/24"    "52.95.255.0/28"    "52.95.255.16/28"  
#> [490] "52.95.255.32/28"   "52.95.255.48/28"   "52.95.255.64/28"  
#> [493] "52.95.255.80/28"   "52.95.255.96/28"   "52.95.255.112/28" 
#> [496] "52.95.255.128/28"  "52.95.255.144/28"  "52.192.0.0/15"    
#> [499] "52.196.0.0/14"     "52.200.0.0/13"     "52.208.0.0/13"    
#> [502] "52.220.0.0/15"     "52.222.0.0/17"     "54.64.0.0/15"     
#> [505] "54.66.0.0/16"      "54.67.0.0/16"      "54.68.0.0/14"     
#> [508] "54.72.0.0/15"      "54.74.0.0/15"      "54.76.0.0/15"     
#> [511] "54.78.0.0/16"      "54.79.0.0/16"      "54.80.0.0/13"     
#> [514] "54.88.0.0/14"      "54.92.0.0/17"      "54.92.128.0/17"   
#> [517] "54.93.0.0/16"      "54.94.0.0/16"      "54.95.0.0/16"     
#> [520] "54.144.0.0/14"     "54.148.0.0/15"     "54.150.0.0/16"    
#> [523] "54.151.0.0/17"     "54.151.128.0/17"   "54.152.0.0/16"    
#> [526] "54.153.0.0/17"     "54.153.128.0/17"   "54.154.0.0/16"    
#> [529] "54.155.0.0/16"     "54.156.0.0/14"     "54.160.0.0/13"    
#> [532] "54.168.0.0/16"     "54.169.0.0/16"     "54.170.0.0/15"    
#> [535] "54.172.0.0/15"     "54.174.0.0/15"     "54.176.0.0/15"    
#> [538] "54.178.0.0/16"     "54.179.0.0/16"     "54.183.0.0/16"    
#> [541] "54.184.0.0/13"     "54.193.0.0/16"     "54.194.0.0/15"    
#> [544] "54.196.0.0/15"     "54.198.0.0/16"     "54.199.0.0/16"    
#> [547] "54.200.0.0/15"     "54.202.0.0/15"     "54.204.0.0/15"    
#> [550] "54.206.0.0/16"     "54.207.0.0/16"     "54.208.0.0/15"    
#> [553] "54.210.0.0/15"     "54.212.0.0/15"     "54.214.0.0/16"    
#> [556] "54.215.0.0/16"     "54.216.0.0/15"     "54.218.0.0/16"    
#> [559] "54.219.0.0/16"     "54.220.0.0/16"     "54.221.0.0/16"    
#> [562] "54.222.128.0/17"   "54.223.0.0/16"     "54.224.0.0/15"    
#> [565] "54.226.0.0/15"     "54.228.0.0/16"     "54.229.0.0/16"    
#> [568] "54.232.0.0/16"     "54.233.0.0/18"     "54.233.64.0/18"   
#> [571] "54.233.128.0/17"   "54.234.0.0/15"     "54.236.0.0/15"    
#> [574] "54.238.0.0/16"     "54.241.0.0/16"     "54.242.0.0/15"    
#> [577] "54.244.0.0/16"     "54.245.0.0/16"     "54.246.0.0/16"    
#> [580] "54.247.0.0/16"     "54.248.0.0/15"     "54.250.0.0/16"    
#> [583] "54.251.0.0/16"     "54.252.0.0/16"     "54.253.0.0/16"    
#> [586] "54.254.0.0/16"     "54.255.0.0/16"     "67.202.0.0/18"    
#> [589] "72.44.32.0/19"     "75.101.128.0/17"   "79.125.0.0/17"    
#> [592] "96.127.0.0/17"     "103.4.8.0/22"      "107.20.0.0/14"    
#> [595] "122.248.192.0/18"  "174.129.0.0/16"    "175.41.128.0/18"  
#> [598] "175.41.192.0/18"   "176.32.64.0/19"    "176.34.0.0/19"    
#> [601] "176.34.32.0/19"    "176.34.64.0/18"    "176.34.128.0/17"  
#> [604] "177.71.128.0/17"   "184.72.0.0/18"     "184.72.64.0/18"   
#> [607] "184.72.128.0/17"   "184.73.0.0/16"     "184.169.128.0/17" 
#> [610] "185.48.120.0/22"   "204.236.128.0/18"  "204.236.192.0/18" 
#> [613] "216.182.224.0/20"  "52.95.110.0/24"    "205.251.192.0/21" 
#> [616] "54.183.255.128/26" "54.228.16.0/26"    "54.232.40.64/26"  
#> [619] "54.241.32.64/26"   "54.243.31.192/26"  "54.244.52.192/26" 
#> [622] "54.245.168.0/26"   "54.248.220.0/26"   "54.250.253.192/26"
#> [625] "54.251.31.128/26"  "54.252.79.128/26"  "54.252.254.192/26"
#> [628] "54.255.254.192/26" "107.23.255.0/26"   "176.34.159.192/26"
#> [631] "177.71.207.128/26" "13.32.0.0/15"      "52.46.0.0/18"     
#> [634] "52.84.0.0/15"      "52.222.128.0/17"   "54.182.0.0/16"    
#> [637] "54.192.0.0/16"     "54.230.0.0/16"     "54.239.128.0/18"  
#> [640] "54.239.192.0/19"   "54.240.128.0/18"   "204.246.164.0/22" 
#> [643] "204.246.168.0/22"  "204.246.174.0/23"  "204.246.176.0/20" 
#> [646] "205.251.192.0/19"  "205.251.249.0/24"  "205.251.250.0/23" 
#> [649] "205.251.252.0/23"  "205.251.254.0/24"  "216.137.32.0/19"  
#>