Azure ranges

Retrieves the official list of Azure cloud network ranges.

azure_ranges()

Value

data.frame with the cloud region and cidr subnet. As of 2016-10-04, these subnets are IPv4 only.

Details

It is unlikely that this list will change in your analysis session, so it is recommended that you cache the results. Future version will automatically cache the results both in-session and on-disk for a period of time.

Examples

ranges <- azure_ranges()

normalize_ipv4(ranges)

#>    [1] "40.112.124.0/24"    "65.52.128.0/19"     "94.245.97.0/24"    
#>    [4] "104.47.169.0/24"    "104.214.240.0/24"   "137.116.192.0/19"  
#>    [7] "168.63.0.0/19"      "168.63.96.0/20"     "168.63.112.16/28"  
#>   [10] "168.63.112.64/26"   "168.63.112.128/25"  "168.63.113.0/24"   
#>   [13] "168.63.114.0/23"    "168.63.116.0/22"    "168.63.120.0/21"   
#>   [16] "193.149.80.0/22"    "213.199.128.0/21"   "213.199.136.0/22"  
#>   [19] "213.199.180.32/28"  "213.199.180.112/28" "213.199.183.0/24"  
#>   [22] "23.97.128.0/17"     "23.98.46.0/24"      "23.100.0.0/20"     
#>   [25] "23.101.64.0/20"     "40.74.0.0/18"       "40.114.128.0/18"   
#>   [28] "104.40.128.0/17"    "104.45.0.0/18"      "104.45.64.0/20"    
#>   [31] "104.46.32.0/19"     "137.117.128.0/17"   "168.61.56.0/25"    
#>   [34] "168.61.56.128/26"   "168.61.56.192/28"   "168.61.56.224/27"  
#>   [37] "168.61.57.0/24"     "168.61.58.0/26"     "168.61.58.96/27"   
#>   [40] "168.61.58.128/25"   "168.61.59.0/24"     "168.61.60.0/22"    
#>   [43] "191.233.64.0/18"    "191.237.232.0/22"   "191.239.200.0/23"  
#>   [46] "191.239.202.0/24"   "191.239.203.0/27"   "191.239.203.64/27" 
#>   [49] "193.149.84.0/22"    "13.69.40.48/28"     "13.69.40.64/26"    
#>   [52] "13.69.40.128/25"    "13.69.42.0/23"      "13.69.44.0/22"     
#>   [55] "13.69.64.0/19"      "13.73.128.16/28"    "13.73.128.32/27"   
#>   [58] "13.73.128.64/26"    "13.73.128.128/25"   "13.73.130.0/23"    
#>   [61] "13.73.132.0/22"     "13.73.168.112/28"   "13.73.168.128/25"  
#>   [64] "13.73.170.0/23"     "13.73.172.0/22"     "13.80.32.16/28"    
#>   [67] "13.80.32.32/27"     "13.80.32.64/26"     "13.80.32.128/25"   
#>   [70] "13.80.34.0/23"      "13.80.36.0/22"      "13.80.89.16/28"    
#>   [73] "13.80.89.32/27"     "13.80.89.64/26"     "13.80.89.128/25"   
#>   [76] "13.80.91.16/28"     "13.80.91.32/27"     "13.80.91.64/26"    
#>   [79] "13.80.91.128/25"    "13.80.92.0/22"      "13.80.160.16/28"   
#>   [82] "13.80.160.64/26"    "13.80.160.128/25"   "13.80.161.0/24"    
#>   [85] "13.80.162.0/23"     "13.80.164.0/22"     "13.80.184.16/28"   
#>   [88] "13.80.184.32/27"    "13.80.184.64/26"    "13.80.184.128/25"  
#>   [91] "13.80.186.0/23"     "13.80.188.0/22"     "13.80.192.0/19"    
#>   [94] "13.80.224.0/21"     "13.80.233.16/28"    "13.80.233.32/27"   
#>   [97] "13.80.233.64/26"    "13.80.233.128/25"   "13.80.234.0/23"    
#>  [100] "13.80.236.0/22"     "13.81.16.16/28"     "13.81.16.64/26"    
#>  [103] "13.81.16.128/25"    "13.81.18.0/23"      "13.81.20.0/22"     
#>  [106] "13.81.72.32/27"     "13.81.72.64/26"     "13.81.72.128/25"   
#>  [109] "13.81.75.0/24"      "13.81.76.0/22"      "13.81.128.0/19"    
#>  [112] "13.81.160.16/28"    "13.81.160.32/27"    "13.81.160.64/26"   
#>  [115] "13.81.160.128/25"   "13.81.162.0/23"     "13.81.164.0/22"    
#>  [118] "13.81.232.16/28"    "13.81.232.32/27"    "13.81.232.64/26"   
#>  [121] "13.81.232.128/25"   "13.81.234.0/23"     "13.81.236.0/22"    
#>  [124] "13.95.128.0/19"     "13.95.160.0/20"     "13.95.176.0/21"    
#>  [127] "13.95.184.0/28"     "13.95.185.0/24"     "13.95.192.0/19"    
#>  [130] "13.95.224.0/20"     "13.95.240.0/26"     "13.95.240.64/27"   
#>  [133] "13.95.240.96/28"    "13.95.241.0/24"     "13.95.248.0/21"    
#>  [136] "40.80.0.0/18"       "40.82.0.0/16"       "40.112.36.128/26"  
#>  [139] "40.112.96.0/20"     "40.112.112.0/21"    "40.112.120.0/22"   
#>  [142] "40.112.125.0/24"    "40.112.126.0/23"    "40.113.96.0/19"    
#>  [145] "40.113.128.0/18"    "40.114.192.0/18"    "40.115.0.0/18"     
#>  [148] "40.118.0.0/17"      "52.178.0.0/19"      "52.178.32.0/20"    
#>  [151] "104.46.16.0/21"     "104.47.128.0/19"    "104.47.160.0/21"   
#>  [154] "104.47.168.0/24"    "104.47.170.0/23"    "104.47.172.0/22"   
#>  [157] "104.47.176.0/20"    "104.214.192.0/19"   "104.214.224.0/20"  
#>  [160] "104.214.241.0/24"   "104.214.242.0/23"   "104.214.244.0/22"  
#>  [163] "104.214.248.0/21"   "13.69.0.0/19"       "13.69.32.0/21"     
#>  [166] "13.69.40.0/27"      "13.69.40.32/28"     "13.69.41.0/24"     
#>  [169] "13.69.48.0/20"      "13.69.96.0/19"      "13.73.128.0/28"    
#>  [172] "13.73.129.0/24"     "13.73.136.0/21"     "13.73.144.0/20"    
#>  [175] "13.73.160.0/21"     "13.73.168.0/26"     "13.73.168.64/27"   
#>  [178] "13.73.168.96/28"    "13.73.169.0/24"     "13.73.176.0/20"    
#>  [181] "13.73.224.0/21"     "13.80.0.0/19"       "13.80.32.0/28"     
#>  [184] "13.80.33.0/24"      "13.80.40.0/21"      "13.80.48.0/20"     
#>  [187] "13.80.64.0/20"      "13.80.80.0/21"      "13.80.88.0/24"     
#>  [190] "13.80.89.0/28"      "13.80.90.0/24"      "13.80.91.0/28"     
#>  [193] "13.80.96.0/19"      "13.80.128.0/19"     "13.80.160.0/28"    
#>  [196] "13.80.160.32/27"    "13.80.168.0/21"     "13.80.176.0/21"    
#>  [199] "13.80.184.0/28"     "13.80.185.0/24"     "13.80.232.0/24"    
#>  [202] "13.80.233.0/28"     "13.80.240.0/20"     "13.81.0.0/20"      
#>  [205] "13.81.16.0/28"      "13.81.16.32/27"     "13.81.24.0/21"     
#>  [208] "13.81.32.0/19"      "13.81.64.0/21"      "13.81.72.0/27"     
#>  [211] "13.81.73.0/24"      "13.81.74.0/24"      "13.81.80.0/20"     
#>  [214] "13.81.96.0/19"      "13.81.160.0/28"     "13.81.161.0/24"    
#>  [217] "13.81.168.0/21"     "13.81.176.0/20"     "13.81.192.0/19"    
#>  [220] "13.81.224.0/21"     "13.81.232.0/28"     "13.81.233.0/24"    
#>  [223] "13.81.240.0/20"     "13.88.200.0/21"     "13.93.0.0/20"      
#>  [226] "13.93.16.0/24"      "13.93.17.0/28"      "13.93.24.0/21"     
#>  [229] "13.93.32.0/20"      "13.93.48.0/21"      "13.93.56.0/24"     
#>  [232] "13.93.57.0/28"      "13.93.64.0/19"      "13.93.96.0/27"     
#>  [235] "13.93.97.0/24"      "13.93.98.0/24"      "13.93.104.0/21"    
#>  [238] "13.93.112.0/20"     "13.94.128.0/19"     "13.94.160.0/28"    
#>  [241] "13.94.161.0/24"     "13.94.168.0/21"     "13.94.176.0/20"    
#>  [244] "13.94.192.0/20"     "13.94.208.0/21"     "13.94.216.0/27"    
#>  [247] "13.94.217.0/24"     "13.94.218.0/24"     "13.94.224.0/19"    
#>  [250] "13.95.0.0/19"       "13.95.64.0/19"      "13.95.96.0/26"     
#>  [253] "13.95.96.64/27"     "13.95.96.96/28"     "13.95.96.128/26"   
#>  [256] "13.95.96.192/28"    "13.95.97.0/24"      "13.95.104.0/21"    
#>  [259] "13.95.112.0/20"     "40.68.0.0/19"       "40.68.32.0/20"     
#>  [262] "40.68.48.0/28"      "40.68.49.0/24"      "40.68.56.0/21"     
#>  [265] "40.68.64.0/19"      "40.68.96.0/21"      "40.68.104.0/28"    
#>  [268] "40.68.105.0/24"     "40.68.112.0/20"     "40.68.128.0/19"    
#>  [271] "40.68.160.0/20"     "40.68.176.0/26"     "40.68.176.64/28"   
#>  [274] "40.68.177.0/24"     "40.68.184.0/21"     "40.68.192.0/19"    
#>  [277] "40.68.224.0/21"     "40.68.232.0/26"     "40.68.232.64/28"   
#>  [280] "40.68.233.0/24"     "40.68.240.0/20"     "52.166.0.0/20"     
#>  [283] "52.166.16.0/21"     "52.166.24.0/28"     "52.166.25.0/24"    
#>  [286] "52.166.32.0/19"     "52.166.64.0/20"     "52.166.80.0/25"    
#>  [289] "52.166.81.0/24"     "52.166.82.0/24"     "52.166.88.0/21"    
#>  [292] "52.166.96.0/19"     "52.166.128.0/20"    "52.166.144.0/21"   
#>  [295] "52.166.152.0/28"    "52.166.153.0/24"    "52.166.160.0/19"   
#>  [298] "52.166.192.0/20"    "52.166.208.0/28"    "52.166.209.0/24"   
#>  [301] "52.166.216.0/21"    "52.166.224.0/19"    "52.174.0.0/21"     
#>  [304] "52.174.8.0/26"      "52.174.9.0/24"      "52.174.10.0/24"    
#>  [307] "52.174.16.0/20"     "52.174.32.0/19"     "52.174.64.0/19"    
#>  [310] "52.174.96.0/20"     "52.174.112.0/28"    "52.174.113.0/24"   
#>  [313] "52.174.120.0/21"    "52.174.128.0/19"    "52.174.160.0/21"   
#>  [316] "52.174.168.0/28"    "52.174.169.0/24"    "52.174.176.0/20"   
#>  [319] "52.174.192.0/19"    "52.174.224.48/28"   "52.174.224.64/26"  
#>  [322] "52.174.224.128/27"  "52.174.225.0/24"    "52.174.232.0/21"   
#>  [325] "52.174.240.0/20"    "52.178.64.0/19"     "52.178.96.0/20"    
#>  [328] "52.178.112.0/21"    "52.178.120.0/28"    "52.178.121.0/24"   
#>  [331] "52.232.0.0/19"      "52.232.32.0/20"     "52.232.48.0/24"    
#>  [334] "52.232.49.0/28"     "52.232.56.0/21"     "52.232.64.0/20"    
#>  [337] "52.232.80.0/21"     "52.232.88.0/28"     "52.232.89.0/24"    
#>  [340] "52.232.96.0/19"     "52.233.128.0/20"    "52.233.144.0/28"   
#>  [343] "52.233.145.0/24"    "52.233.152.0/21"    "52.233.160.0/19"   
#>  [346] "52.233.192.0/21"    "52.236.240.0/26"    "52.236.240.64/28"  
#>  [349] "52.236.242.0/23"    "52.236.248.0/21"    "23.96.0.0/18"      
#>  [352] "23.96.64.0/28"      "23.96.64.64/26"     "23.96.64.128/27"   
#>  [355] "23.96.64.160/28"    "23.96.80.0/20"      "23.96.96.0/19"     
#>  [358] "23.100.16.0/20"     "23.101.128.0/20"    "52.179.25.0/24"    
#>  [361] "52.191.106.0/24"    "104.45.128.0/18"    "104.45.192.0/20"   
#>  [364] "137.116.112.0/20"   "137.117.32.0/19"    "137.117.64.0/18"   
#>  [367] "137.135.64.0/18"    "157.56.176.0/21"    "168.61.32.0/20"    
#>  [370] "168.61.48.0/21"     "168.62.32.0/25"     "168.62.32.128/27"  
#>  [373] "168.62.32.192/26"   "168.62.33.0/24"     "168.62.34.0/26"    
#>  [376] "168.62.34.64/28"    "168.62.34.96/27"    "168.62.34.128/25"  
#>  [379] "168.62.35.0/24"     "168.62.36.0/22"     "168.62.40.0/22"    
#>  [382] "168.62.44.0/23"     "168.62.46.0/24"     "168.62.47.0/27"    
#>  [385] "168.62.47.32/28"    "168.62.47.96/27"    "168.62.47.128/25"  
#>  [388] "168.62.48.0/20"     "168.62.160.0/19"    "23.98.45.0/24"     
#>  [391] "40.76.0.0/18"       "40.121.168.0/28"    "40.121.169.0/24"   
#>  [394] "40.121.176.0/20"    "40.121.192.0/19"    "40.121.224.0/28"   
#>  [397] "40.121.225.0/24"    "40.121.232.0/21"    "40.121.240.0/20"   
#>  [400] "52.168.128.0/20"    "52.168.144.0/21"    "52.168.152.0/28"   
#>  [403] "52.168.153.0/24"    "52.170.128.0/19"    "52.170.160.0/20"   
#>  [406] "104.41.128.0/19"    "138.91.96.0/25"     "138.91.96.128/26"  
#>  [409] "138.91.96.192/28"   "138.91.112.0/20"    "191.234.32.0/19"   
#>  [412] "191.236.0.0/18"     "191.238.0.0/22"     "191.238.4.0/24"    
#>  [415] "191.238.6.0/26"     "191.238.6.64/28"    "191.238.7.0/24"    
#>  [418] "191.238.8.0/21"     "191.238.16.0/20"    "191.238.32.0/19"   
#>  [421] "13.68.161.16/28"    "13.68.161.32/27"    "13.68.161.64/26"   
#>  [424] "13.68.161.128/25"   "13.68.163.48/28"    "13.68.163.64/26"   
#>  [427] "13.68.163.128/25"   "13.72.88.64/26"     "13.72.88.128/25"   
#>  [430] "13.72.91.0/24"      "13.72.92.0/22"      "13.72.120.16/28"   
#>  [433] "13.72.120.32/27"    "13.72.120.64/26"    "13.72.120.128/25"  
#>  [436] "13.72.122.0/23"     "13.72.124.0/22"     "40.76.64.0/18"     
#>  [439] "40.76.128.0/17"     "40.114.0.0/17"      "40.117.128.0/17"   
#>  [442] "40.121.0.0/17"      "40.121.128.0/19"    "40.121.160.0/21"   
#>  [445] "40.121.168.16/28"   "40.121.168.32/27"   "40.121.168.64/26"  
#>  [448] "40.121.168.128/25"  "40.121.170.0/23"    "40.121.172.0/22"   
#>  [451] "40.121.224.16/28"   "40.121.224.32/27"   "40.121.224.64/26"  
#>  [454] "40.121.224.128/25"  "40.121.226.0/23"    "40.121.228.0/22"   
#>  [457] "52.170.176.0/28"    "52.170.177.0/24"    "52.170.184.0/21"   
#>  [460] "52.191.0.0/19"      "52.191.32.0/21"     "104.211.0.0/18"    
#>  [463] "191.237.0.0/19"     "191.237.32.0/22"    "191.237.36.0/23"   
#>  [466] "191.237.38.0/27"    "191.237.39.0/24"    "191.237.40.0/21"   
#>  [469] "191.237.48.0/20"    "191.237.64.0/18"    "13.68.128.0/19"    
#>  [472] "13.68.160.0/24"     "13.68.161.0/28"     "13.68.162.0/24"    
#>  [475] "13.68.163.0/27"     "13.68.163.32/28"    "13.68.165.0/25"    
#>  [478] "13.68.166.0/23"     "13.68.168.0/21"     "13.68.176.0/20"    
#>  [481] "13.68.192.0/18"     "13.72.64.0/20"      "13.72.80.0/21"     
#>  [484] "13.72.88.0/26"      "13.72.89.0/24"      "13.72.90.0/24"     
#>  [487] "13.72.96.0/20"      "13.72.112.0/21"     "13.72.120.0/28"    
#>  [490] "13.72.121.0/24"     "13.82.0.0/16"       "13.90.0.0/19"      
#>  [493] "13.90.32.0/20"      "13.90.48.0/28"      "13.90.49.0/24"     
#>  [496] "13.90.56.0/21"      "13.90.64.0/19"      "13.90.96.0/21"     
#>  [499] "13.90.104.0/28"     "13.90.105.0/24"     "13.90.112.0/20"    
#>  [502] "13.90.128.0/19"     "13.90.160.0/28"     "13.90.161.0/24"    
#>  [505] "13.90.168.0/21"     "13.90.176.0/20"     "13.90.192.0/20"    
#>  [508] "13.90.208.0/21"     "13.90.216.0/28"     "13.90.217.0/24"    
#>  [511] "13.90.224.0/19"     "13.92.0.0/19"       "13.92.32.0/20"     
#>  [514] "13.92.48.0/28"      "13.92.49.0/24"      "13.92.56.0/21"     
#>  [517] "13.92.64.0/19"      "13.92.96.0/21"      "13.92.104.0/28"    
#>  [520] "13.92.105.0/24"     "13.92.112.0/20"     "13.92.128.0/20"    
#>  [523] "13.92.144.0/28"     "13.92.145.0/24"     "13.92.152.0/21"    
#>  [526] "13.92.160.0/19"     "13.92.192.0/21"     "13.92.200.0/27"    
#>  [529] "13.92.201.0/24"     "13.92.202.0/24"     "13.92.208.0/20"    
#>  [532] "13.92.224.0/19"     "40.71.0.0/19"       "40.71.32.0/20"     
#>  [535] "40.71.48.0/28"      "40.71.49.0/24"      "40.71.56.0/21"     
#>  [538] "40.71.64.0/19"      "40.71.96.0/21"      "40.71.104.0/26"    
#>  [541] "40.71.105.0/24"     "40.71.112.0/20"     "40.71.160.0/19"    
#>  [544] "40.71.192.0/19"     "40.71.224.0/20"     "40.71.240.0/27"    
#>  [547] "40.71.248.0/21"     "40.80.144.0/21"     "40.85.160.0/19"    
#>  [550] "40.87.0.0/17"       "40.112.48.0/20"     "40.117.32.0/19"    
#>  [553] "40.117.64.0/18"     "52.168.0.0/19"      "52.168.32.0/21"    
#>  [556] "52.168.40.0/28"     "52.168.41.0/24"     "52.168.48.0/20"    
#>  [559] "52.168.64.0/19"     "52.168.96.0/28"     "52.168.97.0/24"    
#>  [562] "52.168.104.0/21"    "52.168.112.0/20"    "52.168.160.0/20"   
#>  [565] "52.168.176.0/21"    "52.168.184.0/28"    "52.168.185.0/24"   
#>  [568] "52.170.0.0/19"      "52.170.32.0/20"     "52.170.48.0/28"    
#>  [571] "52.170.49.0/24"     "52.170.56.0/21"     "52.170.64.0/19"    
#>  [574] "52.170.96.0/21"     "52.170.104.0/28"    "52.170.105.0/24"   
#>  [577] "52.170.112.0/20"    "52.170.192.0/19"    "52.170.224.0/28"   
#>  [580] "52.170.225.0/24"    "52.170.232.0/21"    "52.170.240.0/20"   
#>  [583] "52.179.0.0/20"      "52.179.16.0/21"     "52.179.24.0/27"    
#>  [586] "52.190.0.0/19"      "52.190.32.0/20"     "52.190.48.48/28"   
#>  [589] "52.190.49.0/24"     "52.190.56.0/21"     "52.191.64.0/19"    
#>  [592] "52.191.96.0/21"     "52.191.104.0/28"    "52.191.105.0/24"   
#>  [595] "52.191.112.0/20"    "52.226.128.0/19"    "52.179.80.0/20"    
#>  [598] "52.179.96.0/19"     "52.186.0.0/28"      "52.186.1.0/24"     
#>  [601] "52.186.8.0/21"      "52.186.16.0/20"     "52.186.32.0/20"    
#>  [604] "52.186.48.0/21"     "52.186.56.0/28"     "52.186.57.0/24"    
#>  [607] "52.186.64.0/19"     "52.186.96.0/20"     "52.186.112.0/26"   
#>  [610] "52.186.112.64/28"   "52.186.113.0/24"    "52.186.114.0/24"   
#>  [613] "52.186.120.0/21"    "52.191.192.0/19"    "52.191.224.0/20"   
#>  [616] "52.191.240.0/28"    "52.191.241.0/24"    "52.191.248.0/21"   
#>  [619] "52.224.0.0/19"      "52.224.32.0/21"     "52.224.40.0/28"    
#>  [622] "52.224.41.0/24"     "52.224.48.0/20"     "52.224.64.0/19"    
#>  [625] "52.224.96.0/28"     "52.224.97.0/24"     "52.224.104.0/21"   
#>  [628] "52.224.112.0/20"    "52.224.128.0/20"    "52.224.144.0/21"   
#>  [631] "52.224.152.0/28"    "52.224.153.0/24"    "52.224.160.0/19"   
#>  [634] "52.224.192.0/20"    "52.224.208.0/28"    "52.224.209.0/24"   
#>  [637] "52.224.216.0/21"    "52.224.224.0/19"    "52.226.0.0/21"     
#>  [640] "52.226.8.0/25"      "52.226.8.128/27"    "52.226.8.160/28"   
#>  [643] "52.226.9.0/24"      "52.226.10.0/24"     "52.226.16.0/20"    
#>  [646] "52.226.32.0/19"     "52.226.64.0/18"     "52.226.160.0/20"   
#>  [649] "52.226.176.0/21"    "52.226.184.0/26"    "52.226.185.0/24"   
#>  [652] "52.234.128.0/19"    "52.234.160.0/20"    "52.234.176.0/25"   
#>  [655] "52.234.176.128/28"  "52.234.178.0/23"    "52.234.184.0/21"   
#>  [658] "52.234.192.0/21"    "52.234.200.0/27"    "52.234.201.0/24"   
#>  [661] "52.234.202.0/24"    "52.234.208.0/20"    "52.234.224.0/19"   
#>  [664] "52.239.152.128/27"  "52.240.0.0/19"      "52.240.32.0/20"    
#>  [667] "52.240.48.0/27"     "52.240.48.32/28"    "52.240.48.64/28"   
#>  [670] "52.240.56.0/22"     "52.245.8.0/28"      "23.100.64.0/21"    
#>  [673] "23.101.144.0/20"    "23.101.32.0/21"     "137.116.0.0/18"    
#>  [676] "137.116.64.0/19"    "137.116.96.0/22"    "191.239.224.0/25"  
#>  [679] "193.149.64.0/21"    "104.46.192.0/20"    "23.102.204.0/22"   
#>  [682] "40.75.0.0/17"       "40.84.0.0/17"       "104.46.0.0/21"     
#>  [685] "104.46.96.0/19"     "104.47.200.0/21"    "104.209.128.0/17"  
#>  [688] "104.210.0.0/20"     "191.236.192.0/18"   "191.237.128.0/19"  
#>  [691] "191.237.160.0/22"   "191.237.164.0/23"   "191.237.168.0/21"  
#>  [694] "191.237.176.0/20"   "104.44.88.32/27"    "104.44.88.96/27"   
#>  [697] "104.44.91.96/27"    "13.68.0.0/18"       "13.68.64.16/28"    
#>  [700] "13.68.64.32/27"     "13.68.64.64/26"     "13.68.64.128/25"   
#>  [703] "13.68.66.0/23"      "13.68.68.0/22"      "13.68.120.96/27"   
#>  [706] "13.68.120.128/25"   "13.68.122.0/23"     "13.68.124.0/22"    
#>  [709] "40.70.0.0/19"       "40.70.32.0/28"      "40.70.33.0/24"     
#>  [712] "40.70.40.0/21"      "40.70.48.0/20"      "40.70.64.0/20"     
#>  [715] "40.70.80.0/21"      "40.79.0.0/21"       "40.79.8.0/27"      
#>  [718] "40.79.8.32/28"      "40.79.8.64/27"      "40.79.8.96/28"     
#>  [721] "40.79.9.0/24"       "40.79.16.0/20"      "40.79.32.0/20"     
#>  [724] "40.79.48.0/27"      "40.79.48.32/28"     "40.79.49.0/24"     
#>  [727] "40.79.56.0/21"      "40.79.64.0/20"      "40.79.80.0/21"     
#>  [730] "40.79.90.0/24"      "40.79.91.0/28"      "40.79.92.0/24"     
#>  [733] "40.79.93.0/28"      "40.79.94.0/24"      "40.79.95.0/28"     
#>  [736] "40.81.0.0/17"       "40.123.16.0/27"     "40.123.16.32/28"   
#>  [739] "40.123.17.0/24"     "40.123.24.0/21"     "40.123.32.0/20"    
#>  [742] "40.123.48.0/21"     "40.123.120.0/21"    "104.208.128.0/27"  
#>  [745] "104.208.128.32/28"  "104.208.129.0/24"   "104.208.136.0/21"  
#>  [748] "104.208.144.0/20"   "104.208.160.0/21"   "104.208.192.0/20"  
#>  [751] "104.208.208.0/24"   "104.208.209.0/28"   "104.208.216.0/21"  
#>  [754] "104.208.224.0/20"   "104.208.240.0/21"   "104.208.248.0/27"  
#>  [757] "13.77.64.0/19"      "13.77.112.192/26"   "13.77.112.176/28"  
#>  [760] "13.77.117.0/24"     "13.77.118.0/23"     "13.77.115.64/26"   
#>  [763] "13.77.115.128/25"   "40.90.138.160/27"   "104.44.93.160/27"  
#>  [766] "13.68.64.0/28"      "13.68.65.0/24"      "13.68.72.0/21"     
#>  [769] "13.68.80.0/20"      "13.68.96.0/20"      "13.68.112.0/21"    
#>  [772] "13.68.120.0/26"     "13.68.120.64/27"    "13.68.121.0/24"    
#>  [775] "13.77.96.0/20"      "13.77.112.0/25"     "13.77.112.128/28"  
#>  [778] "40.70.128.0/19"     "40.70.160.0/20"     "40.70.176.0/28"    
#>  [781] "40.70.177.0/24"     "40.70.184.0/21"     "40.70.192.0/19"    
#>  [784] "40.70.224.0/21"     "40.70.232.240/28"   "40.70.233.0/24"    
#>  [787] "40.70.240.0/20"     "52.167.0.0/19"      "52.167.32.0/28"    
#>  [790] "52.167.33.0/24"     "52.167.40.0/21"     "52.167.48.0/20"    
#>  [793] "52.167.64.0/20"     "52.167.80.0/21"     "52.167.88.16/28"   
#>  [796] "52.167.88.32/27"    "52.167.88.64/26"    "52.167.144.0/20"   
#>  [799] "52.167.160.0/19"    "52.167.240.0/27"    "52.167.248.0/21"   
#>  [802] "52.177.0.0/19"      "52.177.32.0/21"     "52.177.48.0/20"    
#>  [805] "52.177.64.0/19"     "52.177.104.0/21"    "52.177.112.0/20"   
#>  [808] "52.177.128.0/20"    "52.177.144.0/21"    "52.177.160.0/21"   
#>  [811] "52.177.208.0/28"    "52.177.208.64/27"   "52.177.209.0/24"   
#>  [814] "52.177.216.0/21"    "52.177.224.0/19"    "52.179.144.80/28"  
#>  [817] "52.179.146.0/24"    "52.179.192.0/19"    "52.179.224.0/20"   
#>  [820] "52.179.240.32/27"   "52.179.240.64/26"   "52.179.240.128/26" 
#>  [823] "52.179.240.192/27"  "52.184.168.0/28"    "52.184.168.128/28" 
#>  [826] "52.184.169.0/24"    "52.184.170.0/24"    "52.184.176.0/20"   
#>  [829] "52.184.192.0/18"    "52.225.136.0/27"    "52.225.136.32/28"  
#>  [832] "13.77.112.144/28"   "13.77.112.160/28"   "13.77.113.0/24"    
#>  [835] "13.77.114.0/24"     "13.77.115.0/26"     "13.77.116.0/24"    
#>  [838] "13.77.120.0/21"     "52.167.96.0/19"     "52.167.128.0/20"   
#>  [841] "52.167.192.0/19"    "52.167.224.0/20"    "52.177.168.0/21"   
#>  [844] "52.177.176.0/20"    "52.177.192.0/20"    "52.179.128.0/20"   
#>  [847] "52.179.144.16/28"   "52.179.144.32/27"   "52.179.144.64/28"  
#>  [850] "52.179.152.0/21"    "52.179.160.0/19"    "52.179.240.0/27"   
#>  [853] "52.179.240.224/27"  "52.179.241.0/26"    "52.179.241.64/28"  
#>  [856] "52.179.242.0/24"    "52.179.248.0/21"    "52.184.128.0/19"   
#>  [859] "52.184.160.0/21"    "52.184.168.80/28"   "52.184.168.96/27"  
#>  [862] "52.225.128.0/21"    "52.225.136.64/28"   "52.225.137.0/24"   
#>  [865] "52.225.192.0/19"    "52.225.224.0/20"    "52.225.240.0/28"   
#>  [868] "52.225.248.0/21"    "52.232.160.0/19"    "52.232.192.0/19"   
#>  [871] "52.232.224.0/21"    "52.232.232.0/25"    "52.232.232.128/26" 
#>  [874] "52.232.232.192/28"  "52.232.233.0/24"    "52.232.240.0/20"   
#>  [877] "13.83.152.0/28"     "13.83.153.0/24"     "13.83.160.0/19"    
#>  [880] "13.83.192.0/20"     "13.83.208.0/26"     "13.83.208.80/28"   
#>  [883] "13.83.208.128/27"   "13.83.208.160/28"   "13.83.208.192/27"  
#>  [886] "13.83.209.0/24"     "13.83.216.0/21"     "13.83.224.0/19"    
#>  [889] "23.100.32.0/20"     "23.101.192.0/20"    "52.160.136.0/21"   
#>  [892] "52.160.209.0/24"    "104.45.208.0/20"    "104.45.224.0/19"   
#>  [895] "137.116.184.0/21"   "137.117.0.0/19"     "137.135.0.0/18"    
#>  [898] "138.91.64.0/19"     "157.56.160.0/21"    "168.61.0.0/19"     
#>  [901] "168.61.64.0/20"     "168.62.0.0/21"      "168.62.8.0/22"     
#>  [904] "168.62.12.0/23"     "168.62.14.0/24"     "168.62.15.0/27"    
#>  [907] "168.62.15.32/28"    "168.62.15.96/27"    "168.62.15.128/25"  
#>  [910] "168.62.16.0/20"     "168.62.192.0/19"    "168.63.88.0/24"    
#>  [913] "13.83.208.64/28"    "13.83.208.96/27"    "13.83.208.224/28"  
#>  [916] "13.83.210.0/24"     "23.99.64.0/19"      "52.180.64.0/19"    
#>  [919] "52.180.96.0/20"     "138.91.128.48/28"   "138.91.128.64/26"  
#>  [922] "138.91.128.128/25"  "138.91.129.0/26"    "138.91.129.64/28"  
#>  [925] "138.91.136.0/21"    "138.91.144.0/20"    "138.91.160.0/19"   
#>  [928] "138.91.192.0/21"    "138.91.224.0/19"    "191.238.70.0/23"   
#>  [931] "13.83.32.0/24"      "13.83.33.0/28"      "13.83.40.0/21"     
#>  [934] "13.83.48.0/20"      "13.83.64.0/21"      "13.83.72.0/27"     
#>  [937] "13.83.72.32/28"     "13.83.73.0/24"      "13.83.80.0/20"     
#>  [940] "13.83.96.0/20"      "13.83.112.0/24"     "13.83.113.0/28"    
#>  [943] "13.83.120.0/21"     "13.83.128.0/20"     "13.83.144.0/21"    
#>  [946] "23.99.0.0/19"       "23.99.32.0/25"      "23.99.32.128/28"   
#>  [949] "23.99.33.0/28"      "23.99.34.0/26"      "23.99.34.64/27"    
#>  [952] "23.99.34.96/28"     "23.99.34.160/27"    "23.99.34.192/26"   
#>  [955] "23.99.35.0/24"      "23.99.36.0/24"      "23.99.37.48/28"    
#>  [958] "23.99.37.80/28"     "23.99.37.96/27"     "23.99.37.128/27"   
#>  [961] "23.99.37.176/28"    "23.99.37.192/26"    "23.99.38.0/24"     
#>  [964] "23.99.40.0/23"      "23.99.42.0/24"      "23.99.44.0/23"     
#>  [967] "23.99.46.0/24"      "23.99.47.0/26"      "23.99.47.64/28"    
#>  [970] "23.99.48.0/20"      "40.75.128.0/17"     "40.83.152.0/24"    
#>  [973] "40.83.153.0/28"     "40.83.160.0/19"     "40.83.227.0/27"    
#>  [976] "40.112.224.0/26"    "65.52.112.0/20"     "104.40.0.0/18"     
#>  [979] "104.40.64.0/19"     "104.42.0.0/18"      "104.42.64.0/20"    
#>  [982] "104.42.96.0/19"     "104.209.16.0/23"    "104.209.32.0/19"   
#>  [985] "104.210.32.0/19"    "168.63.89.0/25"     "168.63.89.128/26"  
#>  [988] "191.236.64.0/18"    "191.239.0.0/18"     "13.64.0.0/16"      
#>  [991] "13.73.32.0/19"      "13.83.0.0/19"       "13.83.33.16/28"    
#>  [994] "13.83.33.32/27"     "13.83.33.64/26"     "13.83.33.128/25"   
#>  [997] "13.83.34.0/23"      "13.83.36.0/22"      "13.83.72.48/28"    
#> [1000] "13.83.72.64/26"    
#>  [ reached getOption("max.print") -- omitted 1568 entries ]
#>