Digital Ocean ranges

Digitial Ocean (cowardly) doesn't publish their ranges, so we have to approximate with known Digital Ocean ASNs.

digitalocean_ranges()

Value

character vector of IPv4 ranges

Details

It is unlikely that this list will change in your analysis session, so it is recommended that you cache the results. Future version will automatically cache the results both in-session and on-disk for a period of time.

Note

ASN list current as of October, 2016

References

http://bgp.he.net/search?search%5Bsearch%5D=%22Digital+Ocean%22&commit=Search

Examples

ranges <- digitalocean_ranges()

normalize_ipv4(ranges)

#>   [1] "45.55.100.0/22"     "45.55.116.0/22"     "104.131.192.0/19"  
#>   [4] "104.131.224.0/19"   "107.170.0.0/17"     "107.170.20.245/32" 
#>   [7] "107.170.128.0/19"   "107.170.160.0/19"   "138.197.252.0/22"  
#>  [10] "159.203.152.0/22"   "162.243.0.0/17"     "162.243.15.160/32" 
#>  [13] "162.243.36.121/32"  "162.243.37.179/32"  "162.243.112.124/32"
#>  [16] "162.243.191.0/24"   "162.243.192.0/18"   "192.241.160.0/19"  
#>  [19] "192.241.164.234/32" "192.241.164.238/32" "192.241.164.242/32"
#>  [22] "192.241.164.254/32" "192.241.177.47/32"  "192.241.189.28/32" 
#>  [25] "192.241.240.0/20"   "138.197.128.0/20"   "138.197.144.0/20"  
#>  [28] "159.203.0.0/18"     "159.203.52.0/22"    "162.243.190.0/24"  
#>  [31] "45.55.32.0/19"      "45.55.64.0/19"      "45.55.96.0/22"     
#>  [34] "45.55.120.0/22"     "45.55.124.0/22"     "45.55.128.0/18"    
#>  [37] "45.55.192.0/18"     "104.131.0.0/18"     "104.131.64.0/18"   
#>  [40] "104.131.160.0/20"   "104.131.176.0/20"   "104.236.0.0/18"    
#>  [43] "104.236.64.0/18"    "104.236.192.0/18"   "138.197.56.0/22"   
#>  [46] "138.197.60.0/22"    "159.203.64.0/20"    "159.203.80.0/20"   
#>  [49] "159.203.96.0/20"    "159.203.112.0/20"   "159.203.128.0/20"  
#>  [52] "159.203.144.0/22"   "159.203.148.0/22"   "159.203.160.0/20"  
#>  [55] "162.243.176.0/31"   "162.243.176.2/31"   "162.243.176.4/31"  
#>  [58] "162.243.176.6/31"   "162.243.176.8/31"   "162.243.176.10/31" 
#>  [61] "162.243.184.0/22"   "162.243.188.0/23"   "5.101.111.0/24"    
#>  [64] "46.101.0.0/18"      "46.101.64.0/22"     "46.101.72.0/21"    
#>  [67] "46.101.80.0/20"     "138.68.128.0/20"    "138.68.144.0/20"   
#>  [70] "139.59.160.0/20"    "139.59.176.0/20"    "139.59.196.0/22"   
#>  [73] "139.59.200.0/22"    "178.62.0.0/18"      "178.62.64.0/18"    
#>  [76] "188.166.136.0/22"   "188.166.144.0/20"   "188.166.168.0/21"  
#>  [79] "5.101.110.0/24"     "128.199.32.0/19"    "178.62.192.0/18"   
#>  [82] "188.166.0.0/18"     "188.166.64.0/18"    "188.166.132.0/22"  
#>  [85] "188.166.200.0/22"   "5.101.109.0/24"     "46.101.68.0/22"    
#>  [88] "46.101.96.0/20"     "46.101.112.0/20"    "46.101.128.0/18"   
#>  [91] "46.101.192.0/18"    "138.68.64.0/20"     "138.68.80.0/20"    
#>  [94] "138.68.124.0/22"    "139.59.128.0/20"    "139.59.144.0/20"   
#>  [97] "139.59.204.0/22"    "139.59.208.0/21"    "188.166.160.0/21"  
#> [100] "188.166.192.0/22"   "5.101.96.0/21"      "5.101.104.0/22"    
#> [103] "37.139.0.0/19"      "80.240.128.0/20"    "82.196.0.0/20"     
#> [106] "95.85.0.0/18"       "138.68.120.0/23"    "141.0.169.0/24"    
#> [109] "141.0.170.0/24"     "146.185.128.0/19"   "146.185.160.0/20"  
#> [112] "146.185.176.0/21"   "146.185.184.0/21"   "178.62.128.0/18"   
#> [115] "185.14.184.0/22"    "188.166.128.0/22"   "188.166.140.0/22"  
#> [118] "188.226.128.0/17"   "192.81.220.0/22"    "198.199.124.0/22"  
#> [121] "198.211.118.0/23"   "198.211.120.0/21"   "45.55.0.0/19"      
#> [124] "45.55.108.0/22"     "45.55.112.0/22"     "104.131.128.0/20"  
#> [127] "104.131.144.0/20"   "104.236.128.0/18"   "107.170.192.0/18"  
#> [130] "138.68.0.0/20"      "138.68.16.0/20"     "138.68.32.0/24"    
#> [133] "138.68.36.0/22"     "138.68.40.0/21"     "138.68.48.0/20"    
#> [136] "138.68.192.0/22"    "138.68.196.0/22"    "138.68.208.0/20"   
#> [139] "138.197.192.0/20"   "159.203.192.0/20"   "159.203.208.0/20"  
#> [142] "159.203.224.0/20"   "159.203.240.0/20"   "162.243.128.0/19"  
#> [145] "192.241.192.0/19"   "192.241.224.0/20"   "198.199.92.0/22"   
#> [148] "198.199.96.0/20"    "198.199.112.0/21"   "5.101.108.0/24"    
#> [151] "139.59.0.0/17"      "139.59.32.0/20"     "103.253.144.0/22"  
#> [154] "128.199.64.0/18"    "128.199.128.0/18"   "128.199.192.0/18"  
#> [157] "139.59.192.0/22"    "139.59.224.0/20"    "139.59.240.0/20"   
#> [160] "163.47.8.0/22"      "188.166.176.0/20"   "188.166.196.0/22"  
#> [163] "188.166.204.0/22"   "188.166.208.0/20"   "188.166.224.0/20"  
#> [166] "188.166.240.0/20"  
#> attr(,"class")
#> [1] "cidr"         "digitalocean" "character"   
#>