Google Cloud ranges

Retrieves the official list of Google cloud network ranges.

google_ranges()

Value

list of 2 slots, one for ipv4 and one for ipv6 ranges

Details

It is unlikely that this list will change in your analysis session, so it is recommended that you cache the results. Future version will automatically cache the results both in-session and on-disk for a period of time.

Examples

ranges <- google_ranges()

normalize_ipv4(ranges)

#>  [1] "8.34.208.0/20"    "8.35.192.0/21"    "8.35.200.0/23"    "108.59.80.0/20"  
#>  [5] "108.170.192.0/20" "108.170.208.0/21" "108.170.216.0/22" "108.170.220.0/23"
#>  [9] "108.170.222.0/24" "162.216.148.0/22" "162.222.176.0/21" "173.255.112.0/20"
#> [13] "192.158.28.0/22"  "199.192.112.0/22" "199.223.232.0/22" "199.223.236.0/23"
#> [17] "23.236.48.0/20"   "23.251.128.0/19"  "107.167.160.0/19" "107.178.192.0/18"
#> [21] "146.148.2.0/23"   "146.148.4.0/22"   "146.148.8.0/21"   "146.148.16.0/20" 
#> [25] "146.148.32.0/19"  "146.148.64.0/18"  "130.211.4.0/22"   "130.211.8.0/21"  
#> [29] "130.211.16.0/20"  "130.211.32.0/19"  "130.211.64.0/18"  "130.211.128.0/17"
#> [33] "104.154.0.0/15"   "104.196.0.0/14"   "208.68.108.0/23" 
#>