Normalize Cloud CIDR return values

Calling the cloud ranges functions directly returns raw data from the various platform responses. This function will return a flat character vector of IPv4 CIDR blocks from a cloud provider response so it can be more easily used in downstream analysis operations

normalize_ipv4(x)

Arguments

x

the output of one of the cloud _ranges functions

Examples

amazon_ranges() %>% normalize_ipv4()

#>   [1] "13.32.0.0/15"      "13.54.0.0/15"      "23.20.0.0/14"     
#>   [4] "27.0.0.0/22"       "34.192.0.0/12"     "35.154.0.0/16"    
#>   [7] "35.156.0.0/14"     "35.160.0.0/13"     "43.250.192.0/24"  
#>  [10] "43.250.193.0/24"   "46.51.128.0/18"    "46.51.192.0/20"   
#>  [13] "46.51.216.0/21"    "46.51.224.0/19"    "46.137.0.0/17"    
#>  [16] "46.137.128.0/18"   "46.137.192.0/19"   "46.137.224.0/19"  
#>  [19] "50.16.0.0/15"      "50.18.0.0/16"      "50.19.0.0/16"     
#>  [22] "50.112.0.0/16"     "52.0.0.0/15"       "52.2.0.0/15"      
#>  [25] "52.4.0.0/14"       "52.8.0.0/16"       "52.9.0.0/16"      
#>  [28] "52.10.0.0/15"      "52.12.0.0/15"      "52.15.0.0/16"     
#>  [31] "52.16.0.0/15"      "52.18.0.0/15"      "52.20.0.0/14"     
#>  [34] "52.24.0.0/14"      "52.28.0.0/16"      "52.29.0.0/16"     
#>  [37] "52.30.0.0/15"      "52.32.0.0/14"      "52.36.0.0/14"     
#>  [40] "52.40.0.0/14"      "52.44.0.0/15"      "52.46.0.0/18"     
#>  [43] "52.48.0.0/14"      "52.52.0.0/15"      "52.54.0.0/15"     
#>  [46] "52.57.0.0/16"      "52.58.0.0/15"      "52.62.0.0/15"     
#>  [49] "52.64.0.0/17"      "52.64.128.0/17"    "52.65.0.0/16"     
#>  [52] "52.66.0.0/16"      "52.67.0.0/16"      "52.68.0.0/15"     
#>  [55] "52.70.0.0/15"      "52.72.0.0/15"      "52.74.0.0/16"     
#>  [58] "52.76.0.0/17"      "52.76.128.0/17"    "52.77.0.0/16"     
#>  [61] "52.78.0.0/16"      "52.79.0.0/16"      "52.80.0.0/16"     
#>  [64] "52.84.0.0/15"      "52.86.0.0/15"      "52.88.0.0/15"     
#>  [67] "52.90.0.0/15"      "52.92.0.0/20"      "52.92.16.0/20"    
#>  [70] "52.92.32.0/22"     "52.92.39.0/24"     "52.92.40.0/21"    
#>  [73] "52.92.48.0/22"     "52.92.52.0/22"     "52.92.56.0/22"    
#>  [76] "52.92.60.0/22"     "52.92.64.0/22"     "52.92.68.0/22"    
#>  [79] "52.92.72.0/22"     "52.92.76.0/22"     "52.92.80.0/22"    
#>  [82] "52.92.92.0/22"     "52.92.248.0/22"    "52.92.252.0/22"   
#>  [85] "52.93.0.0/24"      "52.93.1.0/24"      "52.93.2.0/24"     
#>  [88] "52.93.3.0/24"      "52.93.4.0/24"      "52.93.8.0/22"     
#>  [91] "52.93.12.0/22"     "52.93.16.0/24"     "52.94.0.0/22"     
#>  [94] "52.94.4.0/24"      "52.94.5.0/24"      "52.94.6.0/24"     
#>  [97] "52.94.7.0/24"      "52.94.8.0/24"      "52.94.9.0/24"     
#> [100] "52.94.10.0/24"     "52.94.11.0/24"     "52.94.12.0/24"    
#> [103] "52.94.13.0/24"     "52.94.17.0/24"     "52.94.24.0/23"    
#> [106] "52.94.28.0/23"     "52.94.192.0/22"    "52.94.196.0/24"   
#> [109] "52.94.197.0/24"    "52.94.198.0/28"    "52.94.198.16/28"  
#> [112] "52.94.198.32/28"   "52.94.198.48/28"   "52.94.198.64/28"  
#> [115] "52.94.198.80/28"   "52.94.198.96/28"   "52.94.198.112/28" 
#> [118] "52.94.199.0/24"    "52.94.200.0/24"    "52.94.204.0/23"   
#> [121] "52.94.206.0/23"    "52.94.208.0/21"    "52.94.216.0/21"   
#> [124] "52.94.224.0/20"    "52.94.252.0/23"    "52.94.254.0/23"   
#> [127] "52.95.0.0/20"      "52.95.16.0/21"     "52.95.24.0/22"    
#> [130] "52.95.28.0/24"     "52.95.30.0/23"     "52.95.34.0/24"    
#> [133] "52.95.35.0/24"     "52.95.36.0/22"     "52.95.40.0/24"    
#> [136] "52.95.48.0/22"     "52.95.52.0/22"     "52.95.56.0/22"    
#> [139] "52.95.60.0/24"     "52.95.61.0/24"     "52.95.62.0/24"    
#> [142] "52.95.63.0/24"     "52.95.64.0/20"     "52.95.80.0/20"    
#> [145] "52.95.96.0/22"     "52.95.100.0/22"    "52.95.104.0/22"   
#> [148] "52.95.110.0/24"    "52.95.128.0/22"    "52.95.132.0/22"   
#> [151] "52.95.136.0/23"    "52.95.138.0/24"    "52.95.142.0/23"   
#> [154] "52.95.144.0/24"    "52.95.192.0/20"    "52.95.212.0/22"   
#> [157] "52.95.240.0/24"    "52.95.241.0/24"    "52.95.242.0/24"   
#> [160] "52.95.243.0/24"    "52.95.244.0/24"    "52.95.245.0/24"   
#> [163] "52.95.246.0/24"    "52.95.247.0/24"    "52.95.248.0/24"   
#> [166] "52.95.249.0/24"    "52.95.251.0/24"    "52.95.252.0/24"   
#> [169] "52.95.255.0/28"    "52.95.255.16/28"   "52.95.255.32/28"  
#> [172] "52.95.255.48/28"   "52.95.255.64/28"   "52.95.255.80/28"  
#> [175] "52.95.255.96/28"   "52.95.255.112/28"  "52.95.255.128/28" 
#> [178] "52.95.255.144/28"  "52.192.0.0/15"     "52.196.0.0/14"    
#> [181] "52.200.0.0/13"     "52.208.0.0/13"     "52.216.0.0/15"    
#> [184] "52.218.0.0/18"     "52.218.64.0/18"    "52.218.128.0/18"  
#> [187] "52.218.192.0/18"   "52.219.0.0/20"     "52.219.16.0/22"   
#> [190] "52.219.20.0/22"    "52.219.24.0/21"    "52.219.32.0/21"   
#> [193] "52.219.40.0/22"    "52.219.44.0/22"    "52.219.56.0/22"   
#> [196] "52.219.60.0/23"    "52.219.62.0/23"    "52.219.64.0/22"   
#> [199] "52.219.68.0/22"    "52.219.72.0/22"    "52.219.76.0/22"   
#> [202] "52.219.80.0/20"    "52.220.0.0/15"     "52.222.0.0/17"    
#> [205] "52.222.128.0/17"   "54.64.0.0/15"      "54.66.0.0/16"     
#> [208] "54.67.0.0/16"      "54.68.0.0/14"      "54.72.0.0/15"     
#> [211] "54.74.0.0/15"      "54.76.0.0/15"      "54.78.0.0/16"     
#> [214] "54.79.0.0/16"      "54.80.0.0/13"      "54.88.0.0/14"     
#> [217] "54.92.0.0/17"      "54.92.128.0/17"    "54.93.0.0/16"     
#> [220] "54.94.0.0/16"      "54.95.0.0/16"      "54.144.0.0/14"    
#> [223] "54.148.0.0/15"     "54.150.0.0/16"     "54.151.0.0/17"    
#> [226] "54.151.128.0/17"   "54.152.0.0/16"     "54.153.0.0/17"    
#> [229] "54.153.128.0/17"   "54.154.0.0/16"     "54.155.0.0/16"    
#> [232] "54.156.0.0/14"     "54.160.0.0/13"     "54.168.0.0/16"    
#> [235] "54.169.0.0/16"     "54.170.0.0/15"     "54.172.0.0/15"    
#> [238] "54.174.0.0/15"     "54.176.0.0/15"     "54.178.0.0/16"    
#> [241] "54.179.0.0/16"     "54.182.0.0/16"     "54.183.0.0/16"    
#> [244] "54.184.0.0/13"     "54.192.0.0/16"     "54.193.0.0/16"    
#> [247] "54.194.0.0/15"     "54.196.0.0/15"     "54.198.0.0/16"    
#> [250] "54.199.0.0/16"     "54.200.0.0/15"     "54.202.0.0/15"    
#> [253] "54.204.0.0/15"     "54.206.0.0/16"     "54.207.0.0/16"    
#> [256] "54.208.0.0/15"     "54.210.0.0/15"     "54.212.0.0/15"    
#> [259] "54.214.0.0/16"     "54.215.0.0/16"     "54.216.0.0/15"    
#> [262] "54.218.0.0/16"     "54.219.0.0/16"     "54.220.0.0/16"    
#> [265] "54.221.0.0/16"     "54.222.0.0/19"     "54.222.58.0/28"   
#> [268] "54.222.128.0/17"   "54.223.0.0/16"     "54.224.0.0/15"    
#> [271] "54.226.0.0/15"     "54.228.0.0/16"     "54.229.0.0/16"    
#> [274] "54.230.0.0/16"     "54.231.0.0/17"     "54.231.128.0/19"  
#> [277] "54.231.160.0/19"   "54.231.192.0/20"   "54.231.224.0/21"  
#> [280] "54.231.232.0/21"   "54.231.240.0/22"   "54.231.244.0/22"  
#> [283] "54.231.248.0/22"   "54.231.252.0/24"   "54.231.253.0/24"  
#> [286] "54.231.254.0/24"   "54.232.0.0/16"     "54.233.0.0/18"    
#> [289] "54.233.64.0/18"    "54.233.128.0/17"   "54.234.0.0/15"    
#> [292] "54.236.0.0/15"     "54.238.0.0/16"     "54.239.2.0/23"    
#> [295] "54.239.4.0/22"     "54.239.8.0/21"     "54.239.16.0/20"   
#> [298] "54.239.32.0/21"    "54.239.48.0/22"    "54.239.52.0/23"   
#> [301] "54.239.54.0/23"    "54.239.56.0/21"    "54.239.64.0/21"   
#> [304] "54.239.96.0/24"    "54.239.98.0/24"    "54.239.99.0/24"   
#> [307] "54.239.100.0/23"   "54.239.104.0/23"   "54.239.108.0/22"  
#> [310] "54.239.114.0/24"   "54.239.116.0/22"   "54.239.120.0/21"  
#> [313] "54.239.128.0/18"   "54.239.192.0/19"   "54.240.128.0/18"  
#> [316] "54.240.192.0/22"   "54.240.196.0/24"   "54.240.197.0/24"  
#> [319] "54.240.198.0/24"   "54.240.199.0/24"   "54.240.200.0/24"  
#> [322] "54.240.202.0/24"   "54.240.203.0/24"   "54.240.204.0/22"  
#> [325] "54.240.208.0/22"   "54.240.212.0/22"   "54.240.216.0/22"  
#> [328] "54.240.220.0/22"   "54.240.225.0/24"   "54.240.226.0/24"  
#> [331] "54.240.227.0/24"   "54.240.228.0/23"   "54.240.230.0/23"  
#> [334] "54.240.232.0/22"   "54.240.236.0/22"   "54.240.240.0/24"  
#> [337] "54.240.244.0/22"   "54.240.248.0/21"   "54.241.0.0/16"    
#> [340] "54.242.0.0/15"     "54.244.0.0/16"     "54.245.0.0/16"    
#> [343] "54.246.0.0/16"     "54.247.0.0/16"     "54.248.0.0/15"    
#> [346] "54.250.0.0/16"     "54.251.0.0/16"     "54.252.0.0/16"    
#> [349] "54.253.0.0/16"     "54.254.0.0/16"     "54.255.0.0/16"    
#> [352] "67.202.0.0/18"     "72.21.192.0/19"    "72.44.32.0/19"    
#> [355] "75.101.128.0/17"   "79.125.0.0/17"     "87.238.80.0/21"   
#> [358] "96.127.0.0/17"     "103.4.8.0/22"      "103.8.172.0/22"   
#> [361] "103.246.148.0/23"  "103.246.150.0/23"  "107.20.0.0/14"    
#> [364] "122.248.192.0/18"  "172.96.97.0/24"    "174.129.0.0/16"   
#> [367] "175.41.128.0/18"   "175.41.192.0/18"   "176.32.64.0/19"   
#> [370] "176.32.96.0/21"    "176.32.104.0/21"   "176.32.112.0/21"  
#> [373] "176.32.120.0/22"   "176.32.125.0/25"   "176.34.0.0/19"    
#> [376] "176.34.32.0/19"    "176.34.64.0/18"    "176.34.128.0/17"  
#> [379] "177.71.128.0/17"   "177.72.240.0/21"   "178.236.0.0/20"   
#> [382] "184.72.0.0/18"     "184.72.64.0/18"    "184.72.128.0/17"  
#> [385] "184.73.0.0/16"     "184.169.128.0/17"  "185.48.120.0/22"  
#> [388] "185.143.16.0/24"   "203.83.220.0/22"   "204.236.128.0/18" 
#> [391] "204.236.192.0/18"  "204.246.160.0/22"  "204.246.164.0/22" 
#> [394] "204.246.168.0/22"  "204.246.174.0/23"  "204.246.176.0/20" 
#> [397] "205.251.192.0/19"  "205.251.224.0/22"  "205.251.228.0/22" 
#> [400] "205.251.232.0/22"  "205.251.236.0/22"  "205.251.240.0/22" 
#> [403] "205.251.244.0/23"  "205.251.247.0/24"  "205.251.248.0/24" 
#> [406] "205.251.249.0/24"  "205.251.250.0/23"  "205.251.252.0/23" 
#> [409] "205.251.254.0/24"  "205.251.255.0/24"  "207.171.160.0/20" 
#> [412] "207.171.176.0/20"  "216.137.32.0/19"   "216.182.224.0/20" 
#> [415] "13.54.0.0/15"      "23.20.0.0/14"      "34.192.0.0/12"    
#> [418] "35.154.0.0/16"     "35.156.0.0/14"     "35.160.0.0/13"    
#> [421] "46.51.128.0/18"    "46.51.192.0/20"    "46.51.216.0/21"   
#> [424] "46.51.224.0/19"    "46.137.0.0/17"     "46.137.128.0/18"  
#> [427] "46.137.192.0/19"   "46.137.224.0/19"   "50.16.0.0/15"     
#> [430] "50.18.0.0/16"      "50.19.0.0/16"      "50.112.0.0/16"    
#> [433] "52.0.0.0/15"       "52.2.0.0/15"       "52.4.0.0/14"      
#> [436] "52.8.0.0/16"       "52.9.0.0/16"       "52.10.0.0/15"     
#> [439] "52.12.0.0/15"      "52.15.0.0/16"      "52.16.0.0/15"     
#> [442] "52.18.0.0/15"      "52.20.0.0/14"      "52.24.0.0/14"     
#> [445] "52.28.0.0/16"      "52.29.0.0/16"      "52.30.0.0/15"     
#> [448] "52.32.0.0/14"      "52.36.0.0/14"      "52.40.0.0/14"     
#> [451] "52.44.0.0/15"      "52.48.0.0/14"      "52.52.0.0/15"     
#> [454] "52.54.0.0/15"      "52.57.0.0/16"      "52.58.0.0/15"     
#> [457] "52.62.0.0/15"      "52.64.0.0/17"      "52.64.128.0/17"   
#> [460] "52.65.0.0/16"      "52.66.0.0/16"      "52.67.0.0/16"     
#> [463] "52.68.0.0/15"      "52.70.0.0/15"      "52.72.0.0/15"     
#> [466] "52.74.0.0/16"      "52.76.0.0/17"      "52.76.128.0/17"   
#> [469] "52.77.0.0/16"      "52.78.0.0/16"      "52.79.0.0/16"     
#> [472] "52.80.0.0/16"      "52.86.0.0/15"      "52.88.0.0/15"     
#> [475] "52.90.0.0/15"      "52.95.240.0/24"    "52.95.241.0/24"   
#> [478] "52.95.242.0/24"    "52.95.243.0/24"    "52.95.244.0/24"   
#> [481] "52.95.245.0/24"    "52.95.246.0/24"    "52.95.247.0/24"   
#> [484] "52.95.248.0/24"    "52.95.249.0/24"    "52.95.251.0/24"   
#> [487] "52.95.252.0/24"    "52.95.255.0/28"    "52.95.255.16/28"  
#> [490] "52.95.255.32/28"   "52.95.255.48/28"   "52.95.255.64/28"  
#> [493] "52.95.255.80/28"   "52.95.255.96/28"   "52.95.255.112/28" 
#> [496] "52.95.255.128/28"  "52.95.255.144/28"  "52.192.0.0/15"    
#> [499] "52.196.0.0/14"     "52.200.0.0/13"     "52.208.0.0/13"    
#> [502] "52.220.0.0/15"     "52.222.0.0/17"     "54.64.0.0/15"     
#> [505] "54.66.0.0/16"      "54.67.0.0/16"      "54.68.0.0/14"     
#> [508] "54.72.0.0/15"      "54.74.0.0/15"      "54.76.0.0/15"     
#> [511] "54.78.0.0/16"      "54.79.0.0/16"      "54.80.0.0/13"     
#> [514] "54.88.0.0/14"      "54.92.0.0/17"      "54.92.128.0/17"   
#> [517] "54.93.0.0/16"      "54.94.0.0/16"      "54.95.0.0/16"     
#> [520] "54.144.0.0/14"     "54.148.0.0/15"     "54.150.0.0/16"    
#> [523] "54.151.0.0/17"     "54.151.128.0/17"   "54.152.0.0/16"    
#> [526] "54.153.0.0/17"     "54.153.128.0/17"   "54.154.0.0/16"    
#> [529] "54.155.0.0/16"     "54.156.0.0/14"     "54.160.0.0/13"    
#> [532] "54.168.0.0/16"     "54.169.0.0/16"     "54.170.0.0/15"    
#> [535] "54.172.0.0/15"     "54.174.0.0/15"     "54.176.0.0/15"    
#> [538] "54.178.0.0/16"     "54.179.0.0/16"     "54.183.0.0/16"    
#> [541] "54.184.0.0/13"     "54.193.0.0/16"     "54.194.0.0/15"    
#> [544] "54.196.0.0/15"     "54.198.0.0/16"     "54.199.0.0/16"    
#> [547] "54.200.0.0/15"     "54.202.0.0/15"     "54.204.0.0/15"    
#> [550] "54.206.0.0/16"     "54.207.0.0/16"     "54.208.0.0/15"    
#> [553] "54.210.0.0/15"     "54.212.0.0/15"     "54.214.0.0/16"    
#> [556] "54.215.0.0/16"     "54.216.0.0/15"     "54.218.0.0/16"    
#> [559] "54.219.0.0/16"     "54.220.0.0/16"     "54.221.0.0/16"    
#> [562] "54.222.128.0/17"   "54.223.0.0/16"     "54.224.0.0/15"    
#> [565] "54.226.0.0/15"     "54.228.0.0/16"     "54.229.0.0/16"    
#> [568] "54.232.0.0/16"     "54.233.0.0/18"     "54.233.64.0/18"   
#> [571] "54.233.128.0/17"   "54.234.0.0/15"     "54.236.0.0/15"    
#> [574] "54.238.0.0/16"     "54.241.0.0/16"     "54.242.0.0/15"    
#> [577] "54.244.0.0/16"     "54.245.0.0/16"     "54.246.0.0/16"    
#> [580] "54.247.0.0/16"     "54.248.0.0/15"     "54.250.0.0/16"    
#> [583] "54.251.0.0/16"     "54.252.0.0/16"     "54.253.0.0/16"    
#> [586] "54.254.0.0/16"     "54.255.0.0/16"     "67.202.0.0/18"    
#> [589] "72.44.32.0/19"     "75.101.128.0/17"   "79.125.0.0/17"    
#> [592] "96.127.0.0/17"     "103.4.8.0/22"      "107.20.0.0/14"    
#> [595] "122.248.192.0/18"  "174.129.0.0/16"    "175.41.128.0/18"  
#> [598] "175.41.192.0/18"   "176.32.64.0/19"    "176.34.0.0/19"    
#> [601] "176.34.32.0/19"    "176.34.64.0/18"    "176.34.128.0/17"  
#> [604] "177.71.128.0/17"   "184.72.0.0/18"     "184.72.64.0/18"   
#> [607] "184.72.128.0/17"   "184.73.0.0/16"     "184.169.128.0/17" 
#> [610] "185.48.120.0/22"   "204.236.128.0/18"  "204.236.192.0/18" 
#> [613] "216.182.224.0/20"  "52.95.110.0/24"    "205.251.192.0/21" 
#> [616] "54.183.255.128/26" "54.228.16.0/26"    "54.232.40.64/26"  
#> [619] "54.241.32.64/26"   "54.243.31.192/26"  "54.244.52.192/26" 
#> [622] "54.245.168.0/26"   "54.248.220.0/26"   "54.250.253.192/26"
#> [625] "54.251.31.128/26"  "54.252.79.128/26"  "54.252.254.192/26"
#> [628] "54.255.254.192/26" "107.23.255.0/26"   "176.34.159.192/26"
#> [631] "177.71.207.128/26" "13.32.0.0/15"      "52.46.0.0/18"     
#> [634] "52.84.0.0/15"      "52.222.128.0/17"   "54.182.0.0/16"    
#> [637] "54.192.0.0/16"     "54.230.0.0/16"     "54.239.128.0/18"  
#> [640] "54.239.192.0/19"   "54.240.128.0/18"   "204.246.164.0/22" 
#> [643] "204.246.168.0/22"  "204.246.174.0/23"  "204.246.176.0/20" 
#> [646] "205.251.192.0/19"  "205.251.249.0/24"  "205.251.250.0/23" 
#> [649] "205.251.252.0/23"  "205.251.254.0/24"  "216.137.32.0/19"  
#> 

azure_ranges() %>% normalize_ipv4()

#>    [1] "40.112.124.0/24"    "65.52.128.0/19"     "94.245.97.0/24"    
#>    [4] "104.47.169.0/24"    "104.214.240.0/24"   "137.116.192.0/19"  
#>    [7] "168.63.0.0/19"      "168.63.96.0/20"     "168.63.112.16/28"  
#>   [10] "168.63.112.64/26"   "168.63.112.128/25"  "168.63.113.0/24"   
#>   [13] "168.63.114.0/23"    "168.63.116.0/22"    "168.63.120.0/21"   
#>   [16] "193.149.80.0/22"    "213.199.128.0/21"   "213.199.136.0/22"  
#>   [19] "213.199.180.32/28"  "213.199.180.112/28" "213.199.183.0/24"  
#>   [22] "23.97.128.0/17"     "23.98.46.0/24"      "23.100.0.0/20"     
#>   [25] "23.101.64.0/20"     "40.74.0.0/18"       "40.114.128.0/18"   
#>   [28] "104.40.128.0/17"    "104.45.0.0/18"      "104.45.64.0/20"    
#>   [31] "104.46.32.0/19"     "137.117.128.0/17"   "168.61.56.0/25"    
#>   [34] "168.61.56.128/26"   "168.61.56.192/28"   "168.61.56.224/27"  
#>   [37] "168.61.57.0/24"     "168.61.58.0/26"     "168.61.58.96/27"   
#>   [40] "168.61.58.128/25"   "168.61.59.0/24"     "168.61.60.0/22"    
#>   [43] "191.233.64.0/18"    "191.237.232.0/22"   "191.239.200.0/23"  
#>   [46] "191.239.202.0/24"   "191.239.203.0/27"   "191.239.203.64/27" 
#>   [49] "193.149.84.0/22"    "13.69.40.48/28"     "13.69.40.64/26"    
#>   [52] "13.69.40.128/25"    "13.69.42.0/23"      "13.69.44.0/22"     
#>   [55] "13.69.64.0/19"      "13.73.128.16/28"    "13.73.128.32/27"   
#>   [58] "13.73.128.64/26"    "13.73.128.128/25"   "13.73.130.0/23"    
#>   [61] "13.73.132.0/22"     "13.73.168.112/28"   "13.73.168.128/25"  
#>   [64] "13.73.170.0/23"     "13.73.172.0/22"     "13.80.32.16/28"    
#>   [67] "13.80.32.32/27"     "13.80.32.64/26"     "13.80.32.128/25"   
#>   [70] "13.80.34.0/23"      "13.80.36.0/22"      "13.80.89.16/28"    
#>   [73] "13.80.89.32/27"     "13.80.89.64/26"     "13.80.89.128/25"   
#>   [76] "13.80.91.16/28"     "13.80.91.32/27"     "13.80.91.64/26"    
#>   [79] "13.80.91.128/25"    "13.80.92.0/22"      "13.80.160.16/28"   
#>   [82] "13.80.160.64/26"    "13.80.160.128/25"   "13.80.161.0/24"    
#>   [85] "13.80.162.0/23"     "13.80.164.0/22"     "13.80.184.16/28"   
#>   [88] "13.80.184.32/27"    "13.80.184.64/26"    "13.80.184.128/25"  
#>   [91] "13.80.186.0/23"     "13.80.188.0/22"     "13.80.192.0/19"    
#>   [94] "13.80.224.0/21"     "13.80.233.16/28"    "13.80.233.32/27"   
#>   [97] "13.80.233.64/26"    "13.80.233.128/25"   "13.80.234.0/23"    
#>  [100] "13.80.236.0/22"     "13.81.16.16/28"     "13.81.16.64/26"    
#>  [103] "13.81.16.128/25"    "13.81.18.0/23"      "13.81.20.0/22"     
#>  [106] "13.81.72.32/27"     "13.81.72.64/26"     "13.81.72.128/25"   
#>  [109] "13.81.75.0/24"      "13.81.76.0/22"      "13.81.128.0/19"    
#>  [112] "13.81.160.16/28"    "13.81.160.32/27"    "13.81.160.64/26"   
#>  [115] "13.81.160.128/25"   "13.81.162.0/23"     "13.81.164.0/22"    
#>  [118] "13.81.232.16/28"    "13.81.232.32/27"    "13.81.232.64/26"   
#>  [121] "13.81.232.128/25"   "13.81.234.0/23"     "13.81.236.0/22"    
#>  [124] "13.95.128.0/19"     "13.95.160.0/20"     "13.95.176.0/21"    
#>  [127] "13.95.184.0/28"     "13.95.185.0/24"     "13.95.192.0/19"    
#>  [130] "13.95.224.0/20"     "13.95.240.0/26"     "13.95.240.64/27"   
#>  [133] "13.95.240.96/28"    "13.95.241.0/24"     "13.95.248.0/21"    
#>  [136] "40.80.0.0/18"       "40.82.0.0/16"       "40.112.36.128/26"  
#>  [139] "40.112.96.0/20"     "40.112.112.0/21"    "40.112.120.0/22"   
#>  [142] "40.112.125.0/24"    "40.112.126.0/23"    "40.113.96.0/19"    
#>  [145] "40.113.128.0/18"    "40.114.192.0/18"    "40.115.0.0/18"     
#>  [148] "40.118.0.0/17"      "52.178.0.0/19"      "52.178.32.0/20"    
#>  [151] "104.46.16.0/21"     "104.47.128.0/19"    "104.47.160.0/21"   
#>  [154] "104.47.168.0/24"    "104.47.170.0/23"    "104.47.172.0/22"   
#>  [157] "104.47.176.0/20"    "104.214.192.0/19"   "104.214.224.0/20"  
#>  [160] "104.214.241.0/24"   "104.214.242.0/23"   "104.214.244.0/22"  
#>  [163] "104.214.248.0/21"   "13.69.0.0/19"       "13.69.32.0/21"     
#>  [166] "13.69.40.0/27"      "13.69.40.32/28"     "13.69.41.0/24"     
#>  [169] "13.69.48.0/20"      "13.69.96.0/19"      "13.73.128.0/28"    
#>  [172] "13.73.129.0/24"     "13.73.136.0/21"     "13.73.144.0/20"    
#>  [175] "13.73.160.0/21"     "13.73.168.0/26"     "13.73.168.64/27"   
#>  [178] "13.73.168.96/28"    "13.73.169.0/24"     "13.73.176.0/20"    
#>  [181] "13.73.224.0/21"     "13.80.0.0/19"       "13.80.32.0/28"     
#>  [184] "13.80.33.0/24"      "13.80.40.0/21"      "13.80.48.0/20"     
#>  [187] "13.80.64.0/20"      "13.80.80.0/21"      "13.80.88.0/24"     
#>  [190] "13.80.89.0/28"      "13.80.90.0/24"      "13.80.91.0/28"     
#>  [193] "13.80.96.0/19"      "13.80.128.0/19"     "13.80.160.0/28"    
#>  [196] "13.80.160.32/27"    "13.80.168.0/21"     "13.80.176.0/21"    
#>  [199] "13.80.184.0/28"     "13.80.185.0/24"     "13.80.232.0/24"    
#>  [202] "13.80.233.0/28"     "13.80.240.0/20"     "13.81.0.0/20"      
#>  [205] "13.81.16.0/28"      "13.81.16.32/27"     "13.81.24.0/21"     
#>  [208] "13.81.32.0/19"      "13.81.64.0/21"      "13.81.72.0/27"     
#>  [211] "13.81.73.0/24"      "13.81.74.0/24"      "13.81.80.0/20"     
#>  [214] "13.81.96.0/19"      "13.81.160.0/28"     "13.81.161.0/24"    
#>  [217] "13.81.168.0/21"     "13.81.176.0/20"     "13.81.192.0/19"    
#>  [220] "13.81.224.0/21"     "13.81.232.0/28"     "13.81.233.0/24"    
#>  [223] "13.81.240.0/20"     "13.88.200.0/21"     "13.93.0.0/20"      
#>  [226] "13.93.16.0/24"      "13.93.17.0/28"      "13.93.24.0/21"     
#>  [229] "13.93.32.0/20"      "13.93.48.0/21"      "13.93.56.0/24"     
#>  [232] "13.93.57.0/28"      "13.93.64.0/19"      "13.93.96.0/27"     
#>  [235] "13.93.97.0/24"      "13.93.98.0/24"      "13.93.104.0/21"    
#>  [238] "13.93.112.0/20"     "13.94.128.0/19"     "13.94.160.0/28"    
#>  [241] "13.94.161.0/24"     "13.94.168.0/21"     "13.94.176.0/20"    
#>  [244] "13.94.192.0/20"     "13.94.208.0/21"     "13.94.216.0/27"    
#>  [247] "13.94.217.0/24"     "13.94.218.0/24"     "13.94.224.0/19"    
#>  [250] "13.95.0.0/19"       "13.95.64.0/19"      "13.95.96.0/26"     
#>  [253] "13.95.96.64/27"     "13.95.96.96/28"     "13.95.96.128/26"   
#>  [256] "13.95.96.192/28"    "13.95.97.0/24"      "13.95.104.0/21"    
#>  [259] "13.95.112.0/20"     "40.68.0.0/19"       "40.68.32.0/20"     
#>  [262] "40.68.48.0/28"      "40.68.49.0/24"      "40.68.56.0/21"     
#>  [265] "40.68.64.0/19"      "40.68.96.0/21"      "40.68.104.0/28"    
#>  [268] "40.68.105.0/24"     "40.68.112.0/20"     "40.68.128.0/19"    
#>  [271] "40.68.160.0/20"     "40.68.176.0/26"     "40.68.176.64/28"   
#>  [274] "40.68.177.0/24"     "40.68.184.0/21"     "40.68.192.0/19"    
#>  [277] "40.68.224.0/21"     "40.68.232.0/26"     "40.68.232.64/28"   
#>  [280] "40.68.233.0/24"     "40.68.240.0/20"     "52.166.0.0/20"     
#>  [283] "52.166.16.0/21"     "52.166.24.0/28"     "52.166.25.0/24"    
#>  [286] "52.166.32.0/19"     "52.166.64.0/20"     "52.166.80.0/25"    
#>  [289] "52.166.81.0/24"     "52.166.82.0/24"     "52.166.88.0/21"    
#>  [292] "52.166.96.0/19"     "52.166.128.0/20"    "52.166.144.0/21"   
#>  [295] "52.166.152.0/28"    "52.166.153.0/24"    "52.166.160.0/19"   
#>  [298] "52.166.192.0/20"    "52.166.208.0/28"    "52.166.209.0/24"   
#>  [301] "52.166.216.0/21"    "52.166.224.0/19"    "52.174.0.0/21"     
#>  [304] "52.174.8.0/26"      "52.174.9.0/24"      "52.174.10.0/24"    
#>  [307] "52.174.16.0/20"     "52.174.32.0/19"     "52.174.64.0/19"    
#>  [310] "52.174.96.0/20"     "52.174.112.0/28"    "52.174.113.0/24"   
#>  [313] "52.174.120.0/21"    "52.174.128.0/19"    "52.174.160.0/21"   
#>  [316] "52.174.168.0/28"    "52.174.169.0/24"    "52.174.176.0/20"   
#>  [319] "52.174.192.0/19"    "52.174.224.48/28"   "52.174.224.64/26"  
#>  [322] "52.174.224.128/27"  "52.174.225.0/24"    "52.174.232.0/21"   
#>  [325] "52.174.240.0/20"    "52.178.64.0/19"     "52.178.96.0/20"    
#>  [328] "52.178.112.0/21"    "52.178.120.0/28"    "52.178.121.0/24"   
#>  [331] "52.232.0.0/19"      "52.232.32.0/20"     "52.232.48.0/24"    
#>  [334] "52.232.49.0/28"     "52.232.56.0/21"     "52.232.64.0/20"    
#>  [337] "52.232.80.0/21"     "52.232.88.0/28"     "52.232.89.0/24"    
#>  [340] "52.232.96.0/19"     "52.233.128.0/20"    "52.233.144.0/28"   
#>  [343] "52.233.145.0/24"    "52.233.152.0/21"    "52.233.160.0/19"   
#>  [346] "52.233.192.0/21"    "52.236.240.0/26"    "52.236.240.64/28"  
#>  [349] "52.236.242.0/23"    "52.236.248.0/21"    "23.96.0.0/18"      
#>  [352] "23.96.64.0/28"      "23.96.64.64/26"     "23.96.64.128/27"   
#>  [355] "23.96.64.160/28"    "23.96.80.0/20"      "23.96.96.0/19"     
#>  [358] "23.100.16.0/20"     "23.101.128.0/20"    "52.179.25.0/24"    
#>  [361] "52.191.106.0/24"    "104.45.128.0/18"    "104.45.192.0/20"   
#>  [364] "137.116.112.0/20"   "137.117.32.0/19"    "137.117.64.0/18"   
#>  [367] "137.135.64.0/18"    "157.56.176.0/21"    "168.61.32.0/20"    
#>  [370] "168.61.48.0/21"     "168.62.32.0/25"     "168.62.32.128/27"  
#>  [373] "168.62.32.192/26"   "168.62.33.0/24"     "168.62.34.0/26"    
#>  [376] "168.62.34.64/28"    "168.62.34.96/27"    "168.62.34.128/25"  
#>  [379] "168.62.35.0/24"     "168.62.36.0/22"     "168.62.40.0/22"    
#>  [382] "168.62.44.0/23"     "168.62.46.0/24"     "168.62.47.0/27"    
#>  [385] "168.62.47.32/28"    "168.62.47.96/27"    "168.62.47.128/25"  
#>  [388] "168.62.48.0/20"     "168.62.160.0/19"    "23.98.45.0/24"     
#>  [391] "40.76.0.0/18"       "40.121.168.0/28"    "40.121.169.0/24"   
#>  [394] "40.121.176.0/20"    "40.121.192.0/19"    "40.121.224.0/28"   
#>  [397] "40.121.225.0/24"    "40.121.232.0/21"    "40.121.240.0/20"   
#>  [400] "52.168.128.0/20"    "52.168.144.0/21"    "52.168.152.0/28"   
#>  [403] "52.168.153.0/24"    "52.170.128.0/19"    "52.170.160.0/20"   
#>  [406] "104.41.128.0/19"    "138.91.96.0/25"     "138.91.96.128/26"  
#>  [409] "138.91.96.192/28"   "138.91.112.0/20"    "191.234.32.0/19"   
#>  [412] "191.236.0.0/18"     "191.238.0.0/22"     "191.238.4.0/24"    
#>  [415] "191.238.6.0/26"     "191.238.6.64/28"    "191.238.7.0/24"    
#>  [418] "191.238.8.0/21"     "191.238.16.0/20"    "191.238.32.0/19"   
#>  [421] "13.68.161.16/28"    "13.68.161.32/27"    "13.68.161.64/26"   
#>  [424] "13.68.161.128/25"   "13.68.163.48/28"    "13.68.163.64/26"   
#>  [427] "13.68.163.128/25"   "13.72.88.64/26"     "13.72.88.128/25"   
#>  [430] "13.72.91.0/24"      "13.72.92.0/22"      "13.72.120.16/28"   
#>  [433] "13.72.120.32/27"    "13.72.120.64/26"    "13.72.120.128/25"  
#>  [436] "13.72.122.0/23"     "13.72.124.0/22"     "40.76.64.0/18"     
#>  [439] "40.76.128.0/17"     "40.114.0.0/17"      "40.117.128.0/17"   
#>  [442] "40.121.0.0/17"      "40.121.128.0/19"    "40.121.160.0/21"   
#>  [445] "40.121.168.16/28"   "40.121.168.32/27"   "40.121.168.64/26"  
#>  [448] "40.121.168.128/25"  "40.121.170.0/23"    "40.121.172.0/22"   
#>  [451] "40.121.224.16/28"   "40.121.224.32/27"   "40.121.224.64/26"  
#>  [454] "40.121.224.128/25"  "40.121.226.0/23"    "40.121.228.0/22"   
#>  [457] "52.170.176.0/28"    "52.170.177.0/24"    "52.170.184.0/21"   
#>  [460] "52.191.0.0/19"      "52.191.32.0/21"     "104.211.0.0/18"    
#>  [463] "191.237.0.0/19"     "191.237.32.0/22"    "191.237.36.0/23"   
#>  [466] "191.237.38.0/27"    "191.237.39.0/24"    "191.237.40.0/21"   
#>  [469] "191.237.48.0/20"    "191.237.64.0/18"    "13.68.128.0/19"    
#>  [472] "13.68.160.0/24"     "13.68.161.0/28"     "13.68.162.0/24"    
#>  [475] "13.68.163.0/27"     "13.68.163.32/28"    "13.68.165.0/25"    
#>  [478] "13.68.166.0/23"     "13.68.168.0/21"     "13.68.176.0/20"    
#>  [481] "13.68.192.0/18"     "13.72.64.0/20"      "13.72.80.0/21"     
#>  [484] "13.72.88.0/26"      "13.72.89.0/24"      "13.72.90.0/24"     
#>  [487] "13.72.96.0/20"      "13.72.112.0/21"     "13.72.120.0/28"    
#>  [490] "13.72.121.0/24"     "13.82.0.0/16"       "13.90.0.0/19"      
#>  [493] "13.90.32.0/20"      "13.90.48.0/28"      "13.90.49.0/24"     
#>  [496] "13.90.56.0/21"      "13.90.64.0/19"      "13.90.96.0/21"     
#>  [499] "13.90.104.0/28"     "13.90.105.0/24"     "13.90.112.0/20"    
#>  [502] "13.90.128.0/19"     "13.90.160.0/28"     "13.90.161.0/24"    
#>  [505] "13.90.168.0/21"     "13.90.176.0/20"     "13.90.192.0/20"    
#>  [508] "13.90.208.0/21"     "13.90.216.0/28"     "13.90.217.0/24"    
#>  [511] "13.90.224.0/19"     "13.92.0.0/19"       "13.92.32.0/20"     
#>  [514] "13.92.48.0/28"      "13.92.49.0/24"      "13.92.56.0/21"     
#>  [517] "13.92.64.0/19"      "13.92.96.0/21"      "13.92.104.0/28"    
#>  [520] "13.92.105.0/24"     "13.92.112.0/20"     "13.92.128.0/20"    
#>  [523] "13.92.144.0/28"     "13.92.145.0/24"     "13.92.152.0/21"    
#>  [526] "13.92.160.0/19"     "13.92.192.0/21"     "13.92.200.0/27"    
#>  [529] "13.92.201.0/24"     "13.92.202.0/24"     "13.92.208.0/20"    
#>  [532] "13.92.224.0/19"     "40.71.0.0/19"       "40.71.32.0/20"     
#>  [535] "40.71.48.0/28"      "40.71.49.0/24"      "40.71.56.0/21"     
#>  [538] "40.71.64.0/19"      "40.71.96.0/21"      "40.71.104.0/26"    
#>  [541] "40.71.105.0/24"     "40.71.112.0/20"     "40.71.160.0/19"    
#>  [544] "40.71.192.0/19"     "40.71.224.0/20"     "40.71.240.0/27"    
#>  [547] "40.71.248.0/21"     "40.80.144.0/21"     "40.85.160.0/19"    
#>  [550] "40.87.0.0/17"       "40.112.48.0/20"     "40.117.32.0/19"    
#>  [553] "40.117.64.0/18"     "52.168.0.0/19"      "52.168.32.0/21"    
#>  [556] "52.168.40.0/28"     "52.168.41.0/24"     "52.168.48.0/20"    
#>  [559] "52.168.64.0/19"     "52.168.96.0/28"     "52.168.97.0/24"    
#>  [562] "52.168.104.0/21"    "52.168.112.0/20"    "52.168.160.0/20"   
#>  [565] "52.168.176.0/21"    "52.168.184.0/28"    "52.168.185.0/24"   
#>  [568] "52.170.0.0/19"      "52.170.32.0/20"     "52.170.48.0/28"    
#>  [571] "52.170.49.0/24"     "52.170.56.0/21"     "52.170.64.0/19"    
#>  [574] "52.170.96.0/21"     "52.170.104.0/28"    "52.170.105.0/24"   
#>  [577] "52.170.112.0/20"    "52.170.192.0/19"    "52.170.224.0/28"   
#>  [580] "52.170.225.0/24"    "52.170.232.0/21"    "52.170.240.0/20"   
#>  [583] "52.179.0.0/20"      "52.179.16.0/21"     "52.179.24.0/27"    
#>  [586] "52.190.0.0/19"      "52.190.32.0/20"     "52.190.48.48/28"   
#>  [589] "52.190.49.0/24"     "52.190.56.0/21"     "52.191.64.0/19"    
#>  [592] "52.191.96.0/21"     "52.191.104.0/28"    "52.191.105.0/24"   
#>  [595] "52.191.112.0/20"    "52.226.128.0/19"    "52.179.80.0/20"    
#>  [598] "52.179.96.0/19"     "52.186.0.0/28"      "52.186.1.0/24"     
#>  [601] "52.186.8.0/21"      "52.186.16.0/20"     "52.186.32.0/20"    
#>  [604] "52.186.48.0/21"     "52.186.56.0/28"     "52.186.57.0/24"    
#>  [607] "52.186.64.0/19"     "52.186.96.0/20"     "52.186.112.0/26"   
#>  [610] "52.186.112.64/28"   "52.186.113.0/24"    "52.186.114.0/24"   
#>  [613] "52.186.120.0/21"    "52.191.192.0/19"    "52.191.224.0/20"   
#>  [616] "52.191.240.0/28"    "52.191.241.0/24"    "52.191.248.0/21"   
#>  [619] "52.224.0.0/19"      "52.224.32.0/21"     "52.224.40.0/28"    
#>  [622] "52.224.41.0/24"     "52.224.48.0/20"     "52.224.64.0/19"    
#>  [625] "52.224.96.0/28"     "52.224.97.0/24"     "52.224.104.0/21"   
#>  [628] "52.224.112.0/20"    "52.224.128.0/20"    "52.224.144.0/21"   
#>  [631] "52.224.152.0/28"    "52.224.153.0/24"    "52.224.160.0/19"   
#>  [634] "52.224.192.0/20"    "52.224.208.0/28"    "52.224.209.0/24"   
#>  [637] "52.224.216.0/21"    "52.224.224.0/19"    "52.226.0.0/21"     
#>  [640] "52.226.8.0/25"      "52.226.8.128/27"    "52.226.8.160/28"   
#>  [643] "52.226.9.0/24"      "52.226.10.0/24"     "52.226.16.0/20"    
#>  [646] "52.226.32.0/19"     "52.226.64.0/18"     "52.226.160.0/20"   
#>  [649] "52.226.176.0/21"    "52.226.184.0/26"    "52.226.185.0/24"   
#>  [652] "52.234.128.0/19"    "52.234.160.0/20"    "52.234.176.0/25"   
#>  [655] "52.234.176.128/28"  "52.234.178.0/23"    "52.234.184.0/21"   
#>  [658] "52.234.192.0/21"    "52.234.200.0/27"    "52.234.201.0/24"   
#>  [661] "52.234.202.0/24"    "52.234.208.0/20"    "52.234.224.0/19"   
#>  [664] "52.239.152.128/27"  "52.240.0.0/19"      "52.240.32.0/20"    
#>  [667] "52.240.48.0/27"     "52.240.48.32/28"    "52.240.48.64/28"   
#>  [670] "52.240.56.0/22"     "52.245.8.0/28"      "23.100.64.0/21"    
#>  [673] "23.101.144.0/20"    "23.101.32.0/21"     "137.116.0.0/18"    
#>  [676] "137.116.64.0/19"    "137.116.96.0/22"    "191.239.224.0/25"  
#>  [679] "193.149.64.0/21"    "104.46.192.0/20"    "23.102.204.0/22"   
#>  [682] "40.75.0.0/17"       "40.84.0.0/17"       "104.46.0.0/21"     
#>  [685] "104.46.96.0/19"     "104.47.200.0/21"    "104.209.128.0/17"  
#>  [688] "104.210.0.0/20"     "191.236.192.0/18"   "191.237.128.0/19"  
#>  [691] "191.237.160.0/22"   "191.237.164.0/23"   "191.237.168.0/21"  
#>  [694] "191.237.176.0/20"   "104.44.88.32/27"    "104.44.88.96/27"   
#>  [697] "104.44.91.96/27"    "13.68.0.0/18"       "13.68.64.16/28"    
#>  [700] "13.68.64.32/27"     "13.68.64.64/26"     "13.68.64.128/25"   
#>  [703] "13.68.66.0/23"      "13.68.68.0/22"      "13.68.120.96/27"   
#>  [706] "13.68.120.128/25"   "13.68.122.0/23"     "13.68.124.0/22"    
#>  [709] "40.70.0.0/19"       "40.70.32.0/28"      "40.70.33.0/24"     
#>  [712] "40.70.40.0/21"      "40.70.48.0/20"      "40.70.64.0/20"     
#>  [715] "40.70.80.0/21"      "40.79.0.0/21"       "40.79.8.0/27"      
#>  [718] "40.79.8.32/28"      "40.79.8.64/27"      "40.79.8.96/28"     
#>  [721] "40.79.9.0/24"       "40.79.16.0/20"      "40.79.32.0/20"     
#>  [724] "40.79.48.0/27"      "40.79.48.32/28"     "40.79.49.0/24"     
#>  [727] "40.79.56.0/21"      "40.79.64.0/20"      "40.79.80.0/21"     
#>  [730] "40.79.90.0/24"      "40.79.91.0/28"      "40.79.92.0/24"     
#>  [733] "40.79.93.0/28"      "40.79.94.0/24"      "40.79.95.0/28"     
#>  [736] "40.81.0.0/17"       "40.123.16.0/27"     "40.123.16.32/28"   
#>  [739] "40.123.17.0/24"     "40.123.24.0/21"     "40.123.32.0/20"    
#>  [742] "40.123.48.0/21"     "40.123.120.0/21"    "104.208.128.0/27"  
#>  [745] "104.208.128.32/28"  "104.208.129.0/24"   "104.208.136.0/21"  
#>  [748] "104.208.144.0/20"   "104.208.160.0/21"   "104.208.192.0/20"  
#>  [751] "104.208.208.0/24"   "104.208.209.0/28"   "104.208.216.0/21"  
#>  [754] "104.208.224.0/20"   "104.208.240.0/21"   "104.208.248.0/27"  
#>  [757] "13.77.64.0/19"      "13.77.112.192/26"   "13.77.112.176/28"  
#>  [760] "13.77.117.0/24"     "13.77.118.0/23"     "13.77.115.64/26"   
#>  [763] "13.77.115.128/25"   "40.90.138.160/27"   "104.44.93.160/27"  
#>  [766] "13.68.64.0/28"      "13.68.65.0/24"      "13.68.72.0/21"     
#>  [769] "13.68.80.0/20"      "13.68.96.0/20"      "13.68.112.0/21"    
#>  [772] "13.68.120.0/26"     "13.68.120.64/27"    "13.68.121.0/24"    
#>  [775] "13.77.96.0/20"      "13.77.112.0/25"     "13.77.112.128/28"  
#>  [778] "40.70.128.0/19"     "40.70.160.0/20"     "40.70.176.0/28"    
#>  [781] "40.70.177.0/24"     "40.70.184.0/21"     "40.70.192.0/19"    
#>  [784] "40.70.224.0/21"     "40.70.232.240/28"   "40.70.233.0/24"    
#>  [787] "40.70.240.0/20"     "52.167.0.0/19"      "52.167.32.0/28"    
#>  [790] "52.167.33.0/24"     "52.167.40.0/21"     "52.167.48.0/20"    
#>  [793] "52.167.64.0/20"     "52.167.80.0/21"     "52.167.88.16/28"   
#>  [796] "52.167.88.32/27"    "52.167.88.64/26"    "52.167.144.0/20"   
#>  [799] "52.167.160.0/19"    "52.167.240.0/27"    "52.167.248.0/21"   
#>  [802] "52.177.0.0/19"      "52.177.32.0/21"     "52.177.48.0/20"    
#>  [805] "52.177.64.0/19"     "52.177.104.0/21"    "52.177.112.0/20"   
#>  [808] "52.177.128.0/20"    "52.177.144.0/21"    "52.177.160.0/21"   
#>  [811] "52.177.208.0/28"    "52.177.208.64/27"   "52.177.209.0/24"   
#>  [814] "52.177.216.0/21"    "52.177.224.0/19"    "52.179.144.80/28"  
#>  [817] "52.179.146.0/24"    "52.179.192.0/19"    "52.179.224.0/20"   
#>  [820] "52.179.240.32/27"   "52.179.240.64/26"   "52.179.240.128/26" 
#>  [823] "52.179.240.192/27"  "52.184.168.0/28"    "52.184.168.128/28" 
#>  [826] "52.184.169.0/24"    "52.184.170.0/24"    "52.184.176.0/20"   
#>  [829] "52.184.192.0/18"    "52.225.136.0/27"    "52.225.136.32/28"  
#>  [832] "13.77.112.144/28"   "13.77.112.160/28"   "13.77.113.0/24"    
#>  [835] "13.77.114.0/24"     "13.77.115.0/26"     "13.77.116.0/24"    
#>  [838] "13.77.120.0/21"     "52.167.96.0/19"     "52.167.128.0/20"   
#>  [841] "52.167.192.0/19"    "52.167.224.0/20"    "52.177.168.0/21"   
#>  [844] "52.177.176.0/20"    "52.177.192.0/20"    "52.179.128.0/20"   
#>  [847] "52.179.144.16/28"   "52.179.144.32/27"   "52.179.144.64/28"  
#>  [850] "52.179.152.0/21"    "52.179.160.0/19"    "52.179.240.0/27"   
#>  [853] "52.179.240.224/27"  "52.179.241.0/26"    "52.179.241.64/28"  
#>  [856] "52.179.242.0/24"    "52.179.248.0/21"    "52.184.128.0/19"   
#>  [859] "52.184.160.0/21"    "52.184.168.80/28"   "52.184.168.96/27"  
#>  [862] "52.225.128.0/21"    "52.225.136.64/28"   "52.225.137.0/24"   
#>  [865] "52.225.192.0/19"    "52.225.224.0/20"    "52.225.240.0/28"   
#>  [868] "52.225.248.0/21"    "52.232.160.0/19"    "52.232.192.0/19"   
#>  [871] "52.232.224.0/21"    "52.232.232.0/25"    "52.232.232.128/26" 
#>  [874] "52.232.232.192/28"  "52.232.233.0/24"    "52.232.240.0/20"   
#>  [877] "13.83.152.0/28"     "13.83.153.0/24"     "13.83.160.0/19"    
#>  [880] "13.83.192.0/20"     "13.83.208.0/26"     "13.83.208.80/28"   
#>  [883] "13.83.208.128/27"   "13.83.208.160/28"   "13.83.208.192/27"  
#>  [886] "13.83.209.0/24"     "13.83.216.0/21"     "13.83.224.0/19"    
#>  [889] "23.100.32.0/20"     "23.101.192.0/20"    "52.160.136.0/21"   
#>  [892] "52.160.209.0/24"    "104.45.208.0/20"    "104.45.224.0/19"   
#>  [895] "137.116.184.0/21"   "137.117.0.0/19"     "137.135.0.0/18"    
#>  [898] "138.91.64.0/19"     "157.56.160.0/21"    "168.61.0.0/19"     
#>  [901] "168.61.64.0/20"     "168.62.0.0/21"      "168.62.8.0/22"     
#>  [904] "168.62.12.0/23"     "168.62.14.0/24"     "168.62.15.0/27"    
#>  [907] "168.62.15.32/28"    "168.62.15.96/27"    "168.62.15.128/25"  
#>  [910] "168.62.16.0/20"     "168.62.192.0/19"    "168.63.88.0/24"    
#>  [913] "13.83.208.64/28"    "13.83.208.96/27"    "13.83.208.224/28"  
#>  [916] "13.83.210.0/24"     "23.99.64.0/19"      "52.180.64.0/19"    
#>  [919] "52.180.96.0/20"     "138.91.128.48/28"   "138.91.128.64/26"  
#>  [922] "138.91.128.128/25"  "138.91.129.0/26"    "138.91.129.64/28"  
#>  [925] "138.91.136.0/21"    "138.91.144.0/20"    "138.91.160.0/19"   
#>  [928] "138.91.192.0/21"    "138.91.224.0/19"    "191.238.70.0/23"   
#>  [931] "13.83.32.0/24"      "13.83.33.0/28"      "13.83.40.0/21"     
#>  [934] "13.83.48.0/20"      "13.83.64.0/21"      "13.83.72.0/27"     
#>  [937] "13.83.72.32/28"     "13.83.73.0/24"      "13.83.80.0/20"     
#>  [940] "13.83.96.0/20"      "13.83.112.0/24"     "13.83.113.0/28"    
#>  [943] "13.83.120.0/21"     "13.83.128.0/20"     "13.83.144.0/21"    
#>  [946] "23.99.0.0/19"       "23.99.32.0/25"      "23.99.32.128/28"   
#>  [949] "23.99.33.0/28"      "23.99.34.0/26"      "23.99.34.64/27"    
#>  [952] "23.99.34.96/28"     "23.99.34.160/27"    "23.99.34.192/26"   
#>  [955] "23.99.35.0/24"      "23.99.36.0/24"      "23.99.37.48/28"    
#>  [958] "23.99.37.80/28"     "23.99.37.96/27"     "23.99.37.128/27"   
#>  [961] "23.99.37.176/28"    "23.99.37.192/26"    "23.99.38.0/24"     
#>  [964] "23.99.40.0/23"      "23.99.42.0/24"      "23.99.44.0/23"     
#>  [967] "23.99.46.0/24"      "23.99.47.0/26"      "23.99.47.64/28"    
#>  [970] "23.99.48.0/20"      "40.75.128.0/17"     "40.83.152.0/24"    
#>  [973] "40.83.153.0/28"     "40.83.160.0/19"     "40.83.227.0/27"    
#>  [976] "40.112.224.0/26"    "65.52.112.0/20"     "104.40.0.0/18"     
#>  [979] "104.40.64.0/19"     "104.42.0.0/18"      "104.42.64.0/20"    
#>  [982] "104.42.96.0/19"     "104.209.16.0/23"    "104.209.32.0/19"   
#>  [985] "104.210.32.0/19"    "168.63.89.0/25"     "168.63.89.128/26"  
#>  [988] "191.236.64.0/18"    "191.239.0.0/18"     "13.64.0.0/16"      
#>  [991] "13.73.32.0/19"      "13.83.0.0/19"       "13.83.33.16/28"    
#>  [994] "13.83.33.32/27"     "13.83.33.64/26"     "13.83.33.128/25"   
#>  [997] "13.83.34.0/23"      "13.83.36.0/22"      "13.83.72.48/28"    
#> [1000] "13.83.72.64/26"    
#>  [ reached getOption("max.print") -- omitted 1568 entries ]
#> 

digitalocean_ranges() %>% normalize_ipv4()

#>   [1] "45.55.100.0/22"     "45.55.116.0/22"     "104.131.192.0/19"  
#>   [4] "104.131.224.0/19"   "107.170.0.0/17"     "107.170.20.245/32" 
#>   [7] "107.170.128.0/19"   "107.170.160.0/19"   "138.197.252.0/22"  
#>  [10] "159.203.152.0/22"   "162.243.0.0/17"     "162.243.15.160/32" 
#>  [13] "162.243.36.121/32"  "162.243.37.179/32"  "162.243.112.124/32"
#>  [16] "162.243.191.0/24"   "162.243.192.0/18"   "192.241.160.0/19"  
#>  [19] "192.241.164.234/32" "192.241.164.238/32" "192.241.164.242/32"
#>  [22] "192.241.164.254/32" "192.241.177.47/32"  "192.241.189.28/32" 
#>  [25] "192.241.240.0/20"   "138.197.128.0/20"   "138.197.144.0/20"  
#>  [28] "159.203.0.0/18"     "159.203.52.0/22"    "162.243.190.0/24"  
#>  [31] "45.55.32.0/19"      "45.55.64.0/19"      "45.55.96.0/22"     
#>  [34] "45.55.120.0/22"     "45.55.124.0/22"     "45.55.128.0/18"    
#>  [37] "45.55.192.0/18"     "104.131.0.0/18"     "104.131.64.0/18"   
#>  [40] "104.131.160.0/20"   "104.131.176.0/20"   "104.236.0.0/18"    
#>  [43] "104.236.64.0/18"    "104.236.192.0/18"   "138.197.56.0/22"   
#>  [46] "138.197.60.0/22"    "159.203.64.0/20"    "159.203.80.0/20"   
#>  [49] "159.203.96.0/20"    "159.203.112.0/20"   "159.203.128.0/20"  
#>  [52] "159.203.144.0/22"   "159.203.148.0/22"   "159.203.160.0/20"  
#>  [55] "162.243.176.0/31"   "162.243.176.2/31"   "162.243.176.4/31"  
#>  [58] "162.243.176.6/31"   "162.243.176.8/31"   "162.243.176.10/31" 
#>  [61] "162.243.184.0/22"   "162.243.188.0/23"   "5.101.111.0/24"    
#>  [64] "46.101.0.0/18"      "46.101.64.0/22"     "46.101.72.0/21"    
#>  [67] "46.101.80.0/20"     "138.68.128.0/20"    "138.68.144.0/20"   
#>  [70] "139.59.160.0/20"    "139.59.176.0/20"    "139.59.196.0/22"   
#>  [73] "139.59.200.0/22"    "178.62.0.0/18"      "178.62.64.0/18"    
#>  [76] "188.166.136.0/22"   "188.166.144.0/20"   "188.166.168.0/21"  
#>  [79] "5.101.110.0/24"     "128.199.32.0/19"    "178.62.192.0/18"   
#>  [82] "188.166.0.0/18"     "188.166.64.0/18"    "188.166.132.0/22"  
#>  [85] "188.166.200.0/22"   "5.101.109.0/24"     "46.101.68.0/22"    
#>  [88] "46.101.96.0/20"     "46.101.112.0/20"    "46.101.128.0/18"   
#>  [91] "46.101.192.0/18"    "138.68.64.0/20"     "138.68.80.0/20"    
#>  [94] "138.68.124.0/22"    "139.59.128.0/20"    "139.59.144.0/20"   
#>  [97] "139.59.204.0/22"    "139.59.208.0/21"    "188.166.160.0/21"  
#> [100] "188.166.192.0/22"   "5.101.96.0/21"      "5.101.104.0/22"    
#> [103] "37.139.0.0/19"      "80.240.128.0/20"    "82.196.0.0/20"     
#> [106] "95.85.0.0/18"       "138.68.120.0/23"    "141.0.169.0/24"    
#> [109] "141.0.170.0/24"     "146.185.128.0/19"   "146.185.160.0/20"  
#> [112] "146.185.176.0/21"   "146.185.184.0/21"   "178.62.128.0/18"   
#> [115] "185.14.184.0/22"    "188.166.128.0/22"   "188.166.140.0/22"  
#> [118] "188.226.128.0/17"   "192.81.220.0/22"    "198.199.124.0/22"  
#> [121] "198.211.118.0/23"   "198.211.120.0/21"   "45.55.0.0/19"      
#> [124] "45.55.108.0/22"     "45.55.112.0/22"     "104.131.128.0/20"  
#> [127] "104.131.144.0/20"   "104.236.128.0/18"   "107.170.192.0/18"  
#> [130] "138.68.0.0/20"      "138.68.16.0/20"     "138.68.32.0/24"    
#> [133] "138.68.36.0/22"     "138.68.40.0/21"     "138.68.48.0/20"    
#> [136] "138.68.192.0/22"    "138.68.196.0/22"    "138.68.208.0/20"   
#> [139] "138.197.192.0/20"   "159.203.192.0/20"   "159.203.208.0/20"  
#> [142] "159.203.224.0/20"   "159.203.240.0/20"   "162.243.128.0/19"  
#> [145] "192.241.192.0/19"   "192.241.224.0/20"   "198.199.92.0/22"   
#> [148] "198.199.96.0/20"    "198.199.112.0/21"   "5.101.108.0/24"    
#> [151] "139.59.0.0/17"      "139.59.32.0/20"     "103.253.144.0/22"  
#> [154] "128.199.64.0/18"    "128.199.128.0/18"   "128.199.192.0/18"  
#> [157] "139.59.192.0/22"    "139.59.224.0/20"    "139.59.240.0/20"   
#> [160] "163.47.8.0/22"      "188.166.176.0/20"   "188.166.196.0/22"  
#> [163] "188.166.204.0/22"   "188.166.208.0/20"   "188.166.224.0/20"  
#> [166] "188.166.240.0/20"  
#> attr(,"class")
#> [1] "cidr"         "digitalocean" "character"   
#> 

google_ranges() %>% normalize_ipv4()

#>  [1] "8.34.208.0/20"    "8.35.192.0/21"    "8.35.200.0/23"    "108.59.80.0/20"  
#>  [5] "108.170.192.0/20" "108.170.208.0/21" "108.170.216.0/22" "108.170.220.0/23"
#>  [9] "108.170.222.0/24" "162.216.148.0/22" "162.222.176.0/21" "173.255.112.0/20"
#> [13] "192.158.28.0/22"  "199.192.112.0/22" "199.223.232.0/22" "199.223.236.0/23"
#> [17] "23.236.48.0/20"   "23.251.128.0/19"  "107.167.160.0/19" "107.178.192.0/18"
#> [21] "146.148.2.0/23"   "146.148.4.0/22"   "146.148.8.0/21"   "146.148.16.0/20" 
#> [25] "146.148.32.0/19"  "146.148.64.0/18"  "130.211.4.0/22"   "130.211.8.0/21"  
#> [29] "130.211.16.0/20"  "130.211.32.0/19"  "130.211.64.0/18"  "130.211.128.0/17"
#> [33] "104.154.0.0/15"   "104.196.0.0/14"   "208.68.108.0/23" 
#> 

rackspace_ranges() %>% normalize_ipv4()

#>   [1] "119.9.0.0/18"     "119.9.128.0/18"   "199.182.60.0/23" 
#>   [4] "199.182.62.0/23"  "103.204.13.0/24"  "119.9.64.0/19"   
#>   [7] "119.9.96.0/19"    "120.136.32.0/20"  "134.8.1.0/24"    
#>  [10] "180.150.128.0/19" "199.166.36.0/24"  "203.60.0.0/17"   
#>  [13] "216.239.98.0/24"  "216.239.100.0/24" "87.225.128.0/17" 
#>  [16] "185.118.172.0/22" "212.21.64.0/19"   "212.42.224.0/19" 
#>  [19] "212.42.225.0/24"  "212.42.236.0/24"  "212.42.237.0/24" 
#>  [22] "23.253.64.0/18"   "23.253.224.0/19"  "50.28.192.0/21"  
#>  [25] "50.28.200.0/22"   "50.28.204.0/22"   "50.28.208.0/22"  
#>  [28] "50.28.212.0/22"   "50.28.216.0/22"   "50.28.220.0/22"  
#>  [31] "50.28.224.0/22"   "50.28.228.0/22"   "50.28.232.0/22"  
#>  [34] "50.28.236.0/22"   "50.56.128.0/17"   "64.39.0.0/19"    
#>  [37] "64.49.192.0/19"   "64.49.224.0/21"   "64.49.234.0/23"  
#>  [40] "64.49.236.0/22"   "64.49.240.0/20"   "65.61.128.0/21"  
#>  [43] "65.61.136.0/22"   "65.61.140.0/24"   "65.61.144.0/23"  
#>  [46] "65.61.148.0/22"   "65.61.153.0/24"   "65.61.154.0/23"  
#>  [49] "65.61.156.0/22"   "65.61.160.0/20"   "65.61.176.0/21"  
#>  [52] "65.61.184.0/23"   "65.61.188.0/22"   "66.216.64.0/19"  
#>  [55] "66.216.96.0/20"   "66.216.112.0/22"  "66.216.118.0/23" 
#>  [58] "66.216.120.0/21"  "67.23.0.0/19"     "67.23.32.0/20"   
#>  [61] "67.192.0.0/16"    "69.20.95.0/24"    "72.3.128.0/17"   
#>  [64] "72.32.0.0/16"     "74.205.32.0/19"   "74.205.64.0/18"  
#>  [67] "98.129.0.0/16"    "104.130.124.0/22" "104.130.128.0/19"
#>  [70] "104.130.240.0/20" "104.239.128.0/19" "104.239.240.0/21"
#>  [73] "108.166.64.0/20"  "108.166.80.0/20"  "108.171.176.0/20"
#>  [76] "148.62.0.0/16"    "162.209.64.0/20"  "162.242.128.0/19"
#>  [79] "166.78.0.0/18"    "166.78.128.0/18"  "172.99.96.0/19"  
#>  [82] "173.45.248.0/21"  "173.203.192.0/18" "174.143.0.0/16"  
#>  [85] "192.237.192.0/19" "198.61.192.0/19"  "198.101.192.0/19"
#>  [88] "198.101.224.0/20" "198.101.240.0/21" "198.101.248.0/21"
#>  [91] "198.254.64.0/21"  "204.63.38.0/24"   "204.63.39.0/24"  
#>  [94] "204.232.192.0/19" "209.61.128.0/19"  "209.61.160.0/22" 
#>  [97] "209.61.164.0/23"  "209.61.168.0/21"  "209.61.176.0/20" 
#> [100] "209.61.184.0/21"  "209.114.32.0/20"  "209.114.48.0/20" 
#> [103] "209.163.232.0/21" "209.163.248.0/22" "209.177.66.0/23" 
#> [106] "216.110.32.0/20"  "23.253.128.0/19"  "65.61.141.0/24"  
#> [109] "65.61.142.0/23"   "65.61.146.0/23"   "65.61.152.0/24"  
#> [112] "65.61.186.0/23"   "67.216.224.0/21"  "67.216.232.0/23" 
#> [115] "67.216.234.0/24"  "67.216.239.0/24"  "69.20.0.0/17"    
#> [118] "72.4.112.0/20"    "74.205.0.0/19"    "104.130.0.0/18"  
#> [121] "104.130.120.0/22" "104.130.160.0/19" "104.130.224.0/20"
#> [124] "104.239.160.0/19" "104.239.224.0/20" "104.239.248.0/21"
#> [127] "142.0.184.0/22"   "146.20.0.0/16"    "156.43.59.0/24"  
#> [130] "162.209.96.0/19"  "162.242.192.0/19" "162.242.224.0/19"
#> [133] "162.247.112.0/21" "162.247.119.0/24" "168.235.224.0/20"
#> [136] "172.99.64.0/19"   "173.203.128.0/18" "198.210.16.0/21" 
#> [139] "201.131.3.0/24"   "204.232.128.0/18" "204.232.224.0/19"
#> [142] "207.97.192.0/18"  "209.0.201.0/24"   "209.87.184.0/22" 
#> [145] "209.133.224.0/19" "216.54.194.0/24"  "50.56.230.0/24"  
#> [148] "23.253.0.0/18"    "23.253.160.0/19"  "23.253.192.0/19" 
#> [151] "50.56.0.0/17"     "50.57.0.0/16"     "64.49.232.0/23"  
#> [154] "65.61.188.0/24"   "66.216.116.0/23"  "67.207.128.0/19" 
#> [157] "74.91.80.0/22"    "74.91.84.0/22"    "74.91.88.0/21"   
#> [160] "104.130.64.0/20"  "104.130.80.0/20"  "104.130.96.0/20" 
#> [163] "104.130.112.0/21" "104.130.192.0/19" "104.239.192.0/19"
#> [166] "108.166.0.0/18"   "108.166.96.0/19"  "108.171.160.0/20"
#> [169] "134.8.2.0/24"     "161.47.0.0/18"    "161.47.64.0/18"  
#> [172] "162.209.0.0/18"   "162.209.80.0/20"  "162.220.0.0/24"  
#> [175] "162.242.160.0/19" "166.78.64.0/18"   "166.78.192.0/18" 
#> [178] "173.45.224.0/20"  "173.45.240.0/21"  "173.203.0.0/17"  
#> [181] "177.152.202.0/24" "184.106.0.0/16"   "185.36.144.0/22" 
#> [184] "192.40.168.0/21"  "192.237.128.0/18" "192.237.224.0/19"
#> [187] "198.61.128.0/18"  "198.61.224.0/19"  "198.101.128.0/18"
#> [190] "198.254.72.0/21"  "199.166.37.0/24"  "208.78.96.0/21"  
#> [193] "209.20.64.0/19"   "209.61.166.0/23"  "216.18.148.0/22" 
#> [196] "5.79.0.0/18"      "31.222.128.0/18"  "37.188.96.0/19"  
#> [199] "46.38.160.0/19"   "78.136.0.0/18"    "83.138.128.0/18" 
#> [202] "89.234.0.0/18"    "91.227.208.0/24"  "92.52.64.0/18"   
#> [205] "94.236.0.0/17"    "95.138.128.0/18"  "134.8.0.0/24"    
#> [208] "134.213.0.0/16"   "142.54.40.0/24"   "146.177.0.0/16"  
#> [211] "162.13.0.0/16"    "164.177.128.0/19" "185.28.196.0/24" 
#> [214] "185.64.72.0/24"   "185.75.60.0/22"   "185.144.228.0/22"
#> [217] "193.142.244.0/24" "212.64.128.0/19"  "212.100.224.0/19"
#> [220] "50.56.228.0/23"   "64.49.247.0/24"  
#> attr(,"class")
#> [1] "cidr"      "rackspace" "character"
#> 

softlayer_ranges() %>% normalize_ipv4()

#>  [1] "159.253.158.0/23" "159.8.198.0/23"   "169.38.118.0/23"  "66.228.118.0/23" 
#>  [5] "173.192.118.0/23" "184.172.118.0/23" "50.22.118.0/23"   "198.23.118.0/23" 
#>  [9] "169.46.118.0/23"  "159.122.118.0/23" "119.81.138.0/23"  "173.193.118.0/23"
#> [13] "5.10.118.0/23"    "168.1.118.0/23"   "169.57.118.0/23"  "159.122.138.0/23"
#> [17] "169.54.118.0/23"  "159.8.118.0/23"   "169.51.118.0/24"  "169.57.138.0/23" 
#> [21] "67.228.118.0/23"  "169.56.118.0/24"  "50.23.118.0/23"   "169.45.118.0/23" 
#> [25] "174.133.118.0/23" "168.1.18.0/23"    "161.202.118.0/23" "158.85.118.0/23" 
#> [29] "208.43.118.0/23"  "169.55.118.0/23" 
#>