Rackspace ranges

Rackspace (cowardly) doesn't publish their ranges, so we have to approximate with known Rackspace Hosting ASNs

rackspace_ranges()

Value

character vector of IPv4 ranges

Details

It is unlikely that this list will change in your analysis session, so it is recommended that you cache the results. Future version will automatically cache the results both in-session and on-disk for a period of time.

Note

ASN list current as of October, 2016

References

http://bgp.he.net/search?search%5Bsearch%5D=%22Rackspace%22&commit=Search

Examples

ranges <- rackspace_ranges()

normalize_ipv4(ranges)

#>   [1] "119.9.0.0/18"     "119.9.128.0/18"   "199.182.60.0/23" 
#>   [4] "199.182.62.0/23"  "103.204.13.0/24"  "119.9.64.0/19"   
#>   [7] "119.9.96.0/19"    "120.136.32.0/20"  "134.8.1.0/24"    
#>  [10] "180.150.128.0/19" "199.166.36.0/24"  "203.60.0.0/17"   
#>  [13] "216.239.98.0/24"  "216.239.100.0/24" "87.225.128.0/17" 
#>  [16] "185.118.172.0/22" "212.21.64.0/19"   "212.42.224.0/19" 
#>  [19] "212.42.225.0/24"  "212.42.236.0/24"  "212.42.237.0/24" 
#>  [22] "23.253.64.0/18"   "23.253.224.0/19"  "50.28.192.0/21"  
#>  [25] "50.28.200.0/22"   "50.28.204.0/22"   "50.28.208.0/22"  
#>  [28] "50.28.212.0/22"   "50.28.216.0/22"   "50.28.220.0/22"  
#>  [31] "50.28.224.0/22"   "50.28.228.0/22"   "50.28.232.0/22"  
#>  [34] "50.28.236.0/22"   "50.56.128.0/17"   "64.39.0.0/19"    
#>  [37] "64.49.192.0/19"   "64.49.224.0/21"   "64.49.234.0/23"  
#>  [40] "64.49.236.0/22"   "64.49.240.0/20"   "65.61.128.0/21"  
#>  [43] "65.61.136.0/22"   "65.61.140.0/24"   "65.61.144.0/23"  
#>  [46] "65.61.148.0/22"   "65.61.153.0/24"   "65.61.154.0/23"  
#>  [49] "65.61.156.0/22"   "65.61.160.0/20"   "65.61.176.0/21"  
#>  [52] "65.61.184.0/23"   "65.61.188.0/22"   "66.216.64.0/19"  
#>  [55] "66.216.96.0/20"   "66.216.112.0/22"  "66.216.118.0/23" 
#>  [58] "66.216.120.0/21"  "67.23.0.0/19"     "67.23.32.0/20"   
#>  [61] "67.192.0.0/16"    "69.20.95.0/24"    "72.3.128.0/17"   
#>  [64] "72.32.0.0/16"     "74.205.32.0/19"   "74.205.64.0/18"  
#>  [67] "98.129.0.0/16"    "104.130.124.0/22" "104.130.128.0/19"
#>  [70] "104.130.240.0/20" "104.239.128.0/19" "104.239.240.0/21"
#>  [73] "108.166.64.0/20"  "108.166.80.0/20"  "108.171.176.0/20"
#>  [76] "148.62.0.0/16"    "162.209.64.0/20"  "162.242.128.0/19"
#>  [79] "166.78.0.0/18"    "166.78.128.0/18"  "172.99.96.0/19"  
#>  [82] "173.45.248.0/21"  "173.203.192.0/18" "174.143.0.0/16"  
#>  [85] "192.237.192.0/19" "198.61.192.0/19"  "198.101.192.0/19"
#>  [88] "198.101.224.0/20" "198.101.240.0/21" "198.101.248.0/21"
#>  [91] "198.254.64.0/21"  "204.63.38.0/24"   "204.63.39.0/24"  
#>  [94] "204.232.192.0/19" "209.61.128.0/19"  "209.61.160.0/22" 
#>  [97] "209.61.164.0/23"  "209.61.168.0/21"  "209.61.176.0/20" 
#> [100] "209.61.184.0/21"  "209.114.32.0/20"  "209.114.48.0/20" 
#> [103] "209.163.232.0/21" "209.163.248.0/22" "209.177.66.0/23" 
#> [106] "216.110.32.0/20"  "23.253.128.0/19"  "65.61.141.0/24"  
#> [109] "65.61.142.0/23"   "65.61.146.0/23"   "65.61.152.0/24"  
#> [112] "65.61.186.0/23"   "67.216.224.0/21"  "67.216.232.0/23" 
#> [115] "67.216.234.0/24"  "67.216.239.0/24"  "69.20.0.0/17"    
#> [118] "72.4.112.0/20"    "74.205.0.0/19"    "104.130.0.0/18"  
#> [121] "104.130.120.0/22" "104.130.160.0/19" "104.130.224.0/20"
#> [124] "104.239.160.0/19" "104.239.224.0/20" "104.239.248.0/21"
#> [127] "142.0.184.0/22"   "146.20.0.0/16"    "156.43.59.0/24"  
#> [130] "162.209.96.0/19"  "162.242.192.0/19" "162.242.224.0/19"
#> [133] "162.247.112.0/21" "162.247.119.0/24" "168.235.224.0/20"
#> [136] "172.99.64.0/19"   "173.203.128.0/18" "198.210.16.0/21" 
#> [139] "201.131.3.0/24"   "204.232.128.0/18" "204.232.224.0/19"
#> [142] "207.97.192.0/18"  "209.0.201.0/24"   "209.87.184.0/22" 
#> [145] "209.133.224.0/19" "216.54.194.0/24"  "50.56.230.0/24"  
#> [148] "23.253.0.0/18"    "23.253.160.0/19"  "23.253.192.0/19" 
#> [151] "50.56.0.0/17"     "50.57.0.0/16"     "64.49.232.0/23"  
#> [154] "65.61.188.0/24"   "66.216.116.0/23"  "67.207.128.0/19" 
#> [157] "74.91.80.0/22"    "74.91.84.0/22"    "74.91.88.0/21"   
#> [160] "104.130.64.0/20"  "104.130.80.0/20"  "104.130.96.0/20" 
#> [163] "104.130.112.0/21" "104.130.192.0/19" "104.239.192.0/19"
#> [166] "108.166.0.0/18"   "108.166.96.0/19"  "108.171.160.0/20"
#> [169] "134.8.2.0/24"     "161.47.0.0/18"    "161.47.64.0/18"  
#> [172] "162.209.0.0/18"   "162.209.80.0/20"  "162.220.0.0/24"  
#> [175] "162.242.160.0/19" "166.78.64.0/18"   "166.78.192.0/18" 
#> [178] "173.45.224.0/20"  "173.45.240.0/21"  "173.203.0.0/17"  
#> [181] "177.152.202.0/24" "184.106.0.0/16"   "185.36.144.0/22" 
#> [184] "192.40.168.0/21"  "192.237.128.0/18" "192.237.224.0/19"
#> [187] "198.61.128.0/18"  "198.61.224.0/19"  "198.101.128.0/18"
#> [190] "198.254.72.0/21"  "199.166.37.0/24"  "208.78.96.0/21"  
#> [193] "209.20.64.0/19"   "209.61.166.0/23"  "216.18.148.0/22" 
#> [196] "5.79.0.0/18"      "31.222.128.0/18"  "37.188.96.0/19"  
#> [199] "46.38.160.0/19"   "78.136.0.0/18"    "83.138.128.0/18" 
#> [202] "89.234.0.0/18"    "91.227.208.0/24"  "92.52.64.0/18"   
#> [205] "94.236.0.0/17"    "95.138.128.0/18"  "134.8.0.0/24"    
#> [208] "134.213.0.0/16"   "142.54.40.0/24"   "146.177.0.0/16"  
#> [211] "162.13.0.0/16"    "164.177.128.0/19" "185.28.196.0/24" 
#> [214] "185.64.72.0/24"   "185.75.60.0/22"   "185.144.228.0/22"
#> [217] "193.142.244.0/24" "212.64.128.0/19"  "212.100.224.0/19"
#> [220] "50.56.228.0/23"   "64.49.247.0/24"  
#> attr(,"class")
#> [1] "cidr"      "rackspace" "character"
#>